+++ src/main/java/TokenUs/TokenUs_BE/web3/contract/VideoNFT.java
@@ -1,0 +1,896 @@
+package TokenUs.TokenUs_BE.web3.contract;
+
+import java.math.BigInteger;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.Callable;
+
+import io.reactivex.Flowable;
+import org.web3j.abi.EventEncoder;
+import org.web3j.abi.FunctionEncoder;
+import org.web3j.abi.TypeReference;
+import org.web3j.abi.datatypes.Address;
+import org.web3j.abi.datatypes.Bool;
+import org.web3j.abi.datatypes.Event;
+import org.web3j.abi.datatypes.Function;
+import org.web3j.abi.datatypes.Type;
+import org.web3j.abi.datatypes.Utf8String;
+import org.web3j.abi.datatypes.generated.Uint256;
+import org.web3j.crypto.Credentials;
+import org.web3j.protocol.Web3j;
+import org.web3j.protocol.core.DefaultBlockParameter;
+import org.web3j.protocol.core.RemoteCall;
+import org.web3j.protocol.core.RemoteFunctionCall;
+import org.web3j.protocol.core.methods.request.EthFilter;
+import org.web3j.protocol.core.methods.response.BaseEventResponse;
+import org.web3j.protocol.core.methods.response.Log;
+import org.web3j.protocol.core.methods.response.TransactionReceipt;
+import org.web3j.tuples.generated.Tuple5;
+import org.web3j.tuples.generated.Tuple6;
+import org.web3j.tx.Contract;
+import org.web3j.tx.TransactionManager;
+import org.web3j.tx.gas.ContractGasProvider;
+
+/**
+ * Auto generated code.
+ *
+ * <p><strong>Do not modify!</strong>
+ *
+ * <p>Please use the <a href="https://docs.web3j.io/command_line.html">web3j command line tools</a>,
+ * or the org.web3j.codegen.SolidityFunctionWrapperGenerator in the <a
+ * href="https://github.com/LFDT-web3j/web3j/tree/main/codegen">codegen module</a> to update.
+ *
+ * <p>Generated with web3j version 1.7.0.
+ */
+@SuppressWarnings("rawtypes")
+public class VideoNFT extends Contract {
+    public static final String BINARY =
+            "0x608060405234801562000010575f80fd5b506040516200463238038062004632833981810160405281019062000036919062000376565b8282815f908162000048919062000644565b5080600190816200005a919062000644565b5050506200007d62000071620000c660201b60201c565b620000cd60201b60201c565b80600f5f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555050505062000728565b5f33905090565b5f600a5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905081600a5f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b5f604051905090565b5f80fd5b5f80fd5b5f80fd5b5f80fd5b5f601f19601f8301169050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52604160045260245ffd5b620001f182620001a9565b810181811067ffffffffffffffff82111715620002135762000212620001b9565b5b80604052505050565b5f6200022762000190565b9050620002358282620001e6565b919050565b5f67ffffffffffffffff821115620002575762000256620001b9565b5b6200026282620001a9565b9050602081019050919050565b5f5b838110156200028e57808201518184015260208101905062000271565b5f8484015250505050565b5f620002af620002a9846200023a565b6200021c565b905082815260208101848484011115620002ce57620002cd620001a5565b5b620002db8482856200026f565b509392505050565b5f82601f830112620002fa57620002f9620001a1565b5b81516200030c84826020860162000299565b91505092915050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f620003408262000315565b9050919050565b620003528162000334565b81146200035d575f80fd5b50565b5f81519050620003708162000347565b92915050565b5f805f6060848603121562000390576200038f62000199565b5b5f84015167ffffffffffffffff811115620003b057620003af6200019d565b5b620003be86828701620002e3565b935050602084015167ffffffffffffffff811115620003e257620003e16200019d565b5b620003f086828701620002e3565b9250506040620004038682870162000360565b9150509250925092565b5f81519050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52602260045260245ffd5b5f60028204905060018216806200045c57607f821691505b60208210810362000472576200047162000417565b5b50919050565b5f819050815f5260205f209050919050565b5f6020601f8301049050919050565b5f82821b905092915050565b5f60088302620004d67fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8262000499565b620004e2868362000499565b95508019841693508086168417925050509392505050565b5f819050919050565b5f819050919050565b5f6200052c620005266200052084620004fa565b62000503565b620004fa565b9050919050565b5f819050919050565b62000547836200050c565b6200055f620005568262000533565b848454620004a5565b825550505050565b5f90565b6200057562000567565b620005828184846200053c565b505050565b5b81811015620005a9576200059d5f826200056b565b60018101905062000588565b5050565b601f821115620005f857620005c28162000478565b620005cd846200048a565b81016020851015620005dd578190505b620005f5620005ec856200048a565b83018262000587565b50505b505050565b5f82821c905092915050565b5f6200061a5f1984600802620005fd565b1980831691505092915050565b5f62000634838362000609565b9150826002028217905092915050565b6200064f826200040d565b67ffffffffffffffff8111156200066b576200066a620001b9565b5b62000677825462000444565b62000684828285620005ad565b5f60209050601f831160018114620006ba575f8415620006a5578287015190505b620006b1858262000627565b86555062000720565b601f198416620006ca8662000478565b5f5b82811015620006f357848901518255600182019150602085019450602081019050620006cc565b868310156200071357848901516200070f601f89168262000609565b8355505b6001600288020188555050505b505050505050565b613efc80620007365f395ff3fe608060405234801561000f575f80fd5b50600436106101a7575f3560e01c806370a08231116100f7578063c87b56dd11610095578063e985e9c51161006f578063e985e9c514610508578063eb26ae7614610538578063f2fde38b1461056c578063fd61827a14610588576101a7565b8063c87b56dd14610485578063daffba8f146104b5578063e6821bf5146104d3576101a7565b806395d89b41116100d157806395d89b4114610413578063a22cb46514610431578063b3cbdc241461044d578063b88d4fde14610469576101a7565b806370a08231146103bb578063715018a6146103eb5780638da5cb5b146103f5576101a7565b806323b872dd1161016457806342842e0e1161013e57806342842e0e1461030f5780634f6ccce71461032b5780636352211e1461035b5780636658166b1461038b576101a7565b806323b872dd146102935780632f745c59146102af578063383b597b146102df576101a7565b806301ffc9a7146101ab57806306fdde03146101db578063081812fc146101f9578063095ea7b31461022957806318160ddd1461024557806320a3dd7514610263575b5f80fd5b6101c560048036038101906101c09190612880565b6105a4565b6040516101d291906128c5565b60405180910390f35b6101e361061d565b6040516101f09190612968565b60405180910390f35b610213600480360381019061020e91906129bb565b6106ac565b6040516102209190612a25565b60405180910390f35b610243600480360381019061023e9190612a68565b6106ee565b005b61024d610804565b60405161025a9190612ab5565b60405180910390f35b61027d600480360381019061027891906129bb565b610810565b60405161028a91906128c5565b60405180910390f35b6102ad60048036038101906102a89190612ace565b61082d565b005b6102c960048036038101906102c49190612a68565b61088d565b6040516102d69190612ab5565b60405180910390f35b6102f960048036038101906102f491906129bb565b61092d565b6040516103069190612ab5565b60405180910390f35b61032960048036038101906103249190612ace565b610947565b005b610345600480360381019061034091906129bb565b610966565b6040516103529190612ab5565b60405180910390f35b610375600480360381019061037091906129bb565b6109d4565b6040516103829190612a25565b60405180910390f35b6103a560048036038101906103a091906129bb565b610a58565b6040516103b29190612ab5565b60405180910390f35b6103d560048036038101906103d09190612b1e565b610a6d565b6040516103e29190612ab5565b60405180910390f35b6103f3610b21565b005b6103fd610b34565b60405161040a9190612a25565b60405180910390f35b61041b610b5c565b6040516104289190612968565b60405180910390f35b61044b60048036038101906104469190612b73565b610bec565b005b61046760048036038101906104629190612cdd565b610c02565b005b610483600480360381019061047e9190612e3c565b610efc565b005b61049f600480360381019061049a91906129bb565b610f5e565b6040516104ac9190612968565b60405180910390f35b6104bd610fc3565b6040516104ca9190612a25565b60405180910390f35b6104ed60048036038101906104e891906129bb565b610fe8565b6040516104ff96959493929190612ebc565b60405180910390f35b610522600480360381019061051d9190612f29565b61114b565b60405161052f91906128c5565b60405180910390f35b610552600480360381019061054d91906129bb565b6111d9565b604051610563959493929190612f67565b60405180910390f35b61058660048036038101906105819190612b1e565b6113bd565b005b6105a2600480360381019061059d9190612b1e565b61143f565b005b5f7f780e9d63000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916148061061657506106158261148a565b5b9050919050565b60605f805461062b90612ff3565b80601f016020809104026020016040519081016040528092919081815260200182805461065790612ff3565b80156106a25780601f10610679576101008083540402835291602001916106a2565b820191905f5260205f20905b81548152906001019060200180831161068557829003601f168201915b5050505050905090565b5f6106b68261156b565b60045f8381526020019081526020015f205f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050919050565b5f6106f8826109d4565b90508073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff1603610768576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161075f90613093565b60405180910390fd5b8073ffffffffffffffffffffffffffffffffffffffff166107876115b6565b73ffffffffffffffffffffffffffffffffffffffff1614806107b657506107b5816107b06115b6565b61114b565b5b6107f5576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107ec90613121565b60405180910390fd5b6107ff83836115bd565b505050565b5f600880549050905090565b600e602052805f5260405f205f915054906101000a900460ff1681565b61083e6108386115b6565b82611673565b61087d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610874906131af565b60405180910390fd5b610888838383611707565b505050565b5f61089783610a6d565b82106108d8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016108cf9061323d565b60405180910390fd5b60065f8473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8381526020019081526020015f2054905092915050565b5f600d5f8381526020019081526020015f20549050919050565b61096183838360405180602001604052805f815250610efc565b505050565b5f61096f610804565b82106109b0576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016109a7906132cb565b60405180910390fd5b600882815481106109c4576109c36132e9565b5b905f5260205f2001549050919050565b5f806109df836119f3565b90505f73ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff1603610a4f576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610a4690613360565b60405180910390fd5b80915050919050565b600d602052805f5260405f205f915090505481565b5f8073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1603610adc576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610ad3906133ee565b60405180910390fd5b60035f8373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f20549050919050565b610b29611a2c565b610b325f611aaa565b565b5f600a5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b606060018054610b6b90612ff3565b80601f0160208091040260200160405190810160405280929190818152602001828054610b9790612ff3565b8015610be25780601f10610bb957610100808354040283529160200191610be2565b820191905f5260205f20905b815481529060010190602001808311610bc557829003601f168201915b5050505050905090565b610bfe610bf76115b6565b8383611b6d565b5050565b600e5f8781526020019081526020015f205f9054906101000a900460ff1615610c60576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610c5790613456565b60405180910390fd5b5f8311610ca2576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610c99906134e4565b60405180910390fd5b5f73ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff1603610d10576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610d079061354c565b60405180910390fd5b6040518060c001604052808781526020018681526020018581526020018273ffffffffffffffffffffffffffffffffffffffff16815260200184815260200183815250600c5f8881526020019081526020015f205f820151815f01556020820151816001019081610d819190613707565b506040820151816002019081610d979190613707565b506060820151816003015f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506080820151816004015560a082015181600501559050506001600e5f8881526020019081526020015f205f6101000a81548160ff0219169083151502179055505f5b83811015610e9e57610e32600b611cd4565b5f610e3d600b611ce8565b9050610e498382611cf4565b87600d5f8381526020019081526020015f2081905550610e8a600f5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff16826115bd565b508080610e9690613803565b915050610e20565b508073ffffffffffffffffffffffffffffffffffffffff16867f5c60f10ddf7c8d48ceeea4dcc0f8a2d8a3c5dd589c2889ac54db10c21c36a06785888887604051610eec949392919061384a565b60405180910390a3505050505050565b610f0d610f076115b6565b83611673565b610f4c576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610f43906131af565b60405180910390fd5b610f5884848484611f07565b50505050565b6060610f698261156b565b5f610f72611f63565b90505f815111610f905760405180602001604052805f815250610fbb565b80610f9a84611f79565b604051602001610fab9291906138d5565b6040516020818303038152906040525b915050919050565b600f5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b600c602052805f5260405f205f91509050805f01549080600101805461100d90612ff3565b80601f016020809104026020016040519081016040528092919081815260200182805461103990612ff3565b80156110845780601f1061105b57610100808354040283529160200191611084565b820191905f5260205f20905b81548152906001019060200180831161106757829003601f168201915b50505050509080600201805461109990612ff3565b80601f01602080910402602001604051908101604052809291908181526020018280546110c590612ff3565b80156111105780601f106110e757610100808354040283529160200191611110565b820191905f5260205f20905b8154815290600101906020018083116110f357829003601f168201915b505050505090806003015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff16908060040154908060050154905086565b5f60055f8473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f9054906101000a900460ff16905092915050565b6060805f805f80600c5f8881526020019081526020015f206040518060c00160405290815f820154815260200160018201805461121590612ff3565b80601f016020809104026020016040519081016040528092919081815260200182805461124190612ff3565b801561128c5780601f106112635761010080835404028352916020019161128c565b820191905f5260205f20905b81548152906001019060200180831161126f57829003601f168201915b505050505081526020016002820180546112a590612ff3565b80601f01602080910402602001604051908101604052809291908181526020018280546112d190612ff3565b801561131c5780601f106112f35761010080835404028352916020019161131c565b820191905f5260205f20905b8154815290600101906020018083116112ff57829003601f168201915b50505050508152602001600382015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200160048201548152602001600582015481525050905080602001518160400151826060015183608001518460a00151955095509550955095505091939590929450565b6113c5611a2c565b5f73ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff1603611433576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161142a90613968565b60405180910390fd5b61143c81611aaa565b50565b611447611a2c565b80600f5f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555050565b5f7f80ac58cd000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916148061155457507f5b5e139f000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916145b80611564575061156382612043565b5b9050919050565b611574816120ac565b6115b3576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016115aa90613360565b60405180910390fd5b50565b5f33905090565b8160045f8381526020019081526020015f205f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550808273ffffffffffffffffffffffffffffffffffffffff1661162d836109d4565b73ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b92560405160405180910390a45050565b5f8061167e836109d4565b90508073ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff1614806116c057506116bf818561114b565b5b806116fe57508373ffffffffffffffffffffffffffffffffffffffff166116e6846106ac565b73ffffffffffffffffffffffffffffffffffffffff16145b91505092915050565b8273ffffffffffffffffffffffffffffffffffffffff16611727826109d4565b73ffffffffffffffffffffffffffffffffffffffff161461177d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611774906139f6565b60405180910390fd5b5f73ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16036117eb576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016117e290613a84565b60405180910390fd5b6117f883838360016120ec565b8273ffffffffffffffffffffffffffffffffffffffff16611818826109d4565b73ffffffffffffffffffffffffffffffffffffffff161461186e576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611865906139f6565b60405180910390fd5b60045f8281526020019081526020015f205f6101000a81549073ffffffffffffffffffffffffffffffffffffffff0219169055600160035f8573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8282540392505081905550600160035f8473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f82825401925050819055508160025f8381526020019081526020015f205f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550808273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef60405160405180910390a46119ee8383836001612247565b505050565b5f60025f8381526020019081526020015f205f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050919050565b611a346115b6565b73ffffffffffffffffffffffffffffffffffffffff16611a52610b34565b73ffffffffffffffffffffffffffffffffffffffff1614611aa8576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611a9f90613aec565b60405180910390fd5b565b5f600a5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905081600a5f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b8173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff1603611bdb576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611bd290613b54565b60405180910390fd5b8060055f8573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f6101000a81548160ff0219169083151502179055508173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167f17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c3183604051611cc791906128c5565b60405180910390a3505050565b6001815f015f828254019250508190555050565b5f815f01549050919050565b5f73ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1603611d62576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611d5990613bbc565b60405180910390fd5b611d6b816120ac565b15611dab576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611da290613c24565b60405180910390fd5b611db85f838360016120ec565b611dc1816120ac565b15611e01576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611df890613c24565b60405180910390fd5b600160035f8473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f82825401925050819055508160025f8381526020019081526020015f205f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550808273ffffffffffffffffffffffffffffffffffffffff165f73ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef60405160405180910390a4611f035f83836001612247565b5050565b611f12848484611707565b611f1e8484848461224d565b611f5d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611f5490613cb2565b60405180910390fd5b50505050565b606060405180602001604052805f815250905090565b60605f6001611f87846123cf565b0190505f8167ffffffffffffffff811115611fa557611fa4612bb9565b5b6040519080825280601f01601f191660200182016040528015611fd75781602001600182028036833780820191505090505b5090505f82602001820190505b600115612038578080600190039150507f3031323334353637383961626364656600000000000000000000000000000000600a86061a8153600a858161202d5761202c613cd0565b5b0494505f8503611fe4575b819350505050919050565b5f7f01ffc9a7000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916149050919050565b5f8073ffffffffffffffffffffffffffffffffffffffff166120cd836119f3565b73ffffffffffffffffffffffffffffffffffffffff1614159050919050565b6120f884848484612520565b600181111561213c576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161213390613d6d565b60405180910390fd5b5f8290505f73ffffffffffffffffffffffffffffffffffffffff168573ffffffffffffffffffffffffffffffffffffffff16036121815761217c81612526565b6121c0565b8373ffffffffffffffffffffffffffffffffffffffff168573ffffffffffffffffffffffffffffffffffffffff16146121bf576121be858261256a565b5b5b5f73ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff1603612201576121fc816126c0565b612240565b8473ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff161461223f5761223e8482612780565b5b5b5050505050565b50505050565b5f61226d8473ffffffffffffffffffffffffffffffffffffffff166127f8565b156123c2578373ffffffffffffffffffffffffffffffffffffffff1663150b7a026122966115b6565b8786866040518563ffffffff1660e01b81526004016122b89493929190613ddd565b6020604051808303815f875af19250505080156122f357506040513d601f19601f820116820180604052508101906122f09190613e3b565b60015b612372573d805f8114612321576040519150601f19603f3d011682016040523d82523d5f602084013e612326565b606091505b505f81510361236a576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161236190613cb2565b60405180910390fd5b805181602001fd5b63150b7a0260e01b7bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916817bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916149150506123c7565b600190505b949350505050565b5f805f90507a184f03e93ff9f4daa797ed6e38ed64bf6a1f010000000000000000831061242b577a184f03e93ff9f4daa797ed6e38ed64bf6a1f010000000000000000838161242157612420613cd0565b5b0492506040810190505b6d04ee2d6d415b85acef81000000008310612468576d04ee2d6d415b85acef8100000000838161245e5761245d613cd0565b5b0492506020810190505b662386f26fc10000831061249757662386f26fc10000838161248d5761248c613cd0565b5b0492506010810190505b6305f5e10083106124c0576305f5e10083816124b6576124b5613cd0565b5b0492506008810190505b61271083106124e55761271083816124db576124da613cd0565b5b0492506004810190505b6064831061250857606483816124fe576124fd613cd0565b5b0492506002810190505b600a8310612517576001810190505b80915050919050565b50505050565b60088054905060095f8381526020019081526020015f2081905550600881908060018154018082558091505060019003905f5260205f20015f909190919091505550565b5f600161257684610a6d565b6125809190613e66565b90505f60075f8481526020019081526020015f20549050818114612657575f60065f8673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8481526020019081526020015f205490508060065f8773ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8481526020019081526020015f20819055508160075f8381526020019081526020015f2081905550505b60075f8481526020019081526020015f205f905560065f8573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8381526020019081526020015f205f905550505050565b5f60016008805490506126d39190613e66565b90505f60095f8481526020019081526020015f205490505f600883815481106126ff576126fe6132e9565b5b905f5260205f2001549050806008838154811061271f5761271e6132e9565b5b905f5260205f2001819055508160095f8381526020019081526020015f208190555060095f8581526020019081526020015f205f9055600880548061276757612766613e99565b5b600190038181905f5260205f20015f9055905550505050565b5f61278a83610a6d565b90508160065f8573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f8381526020019081526020015f20819055508060075f8481526020019081526020015f2081905550505050565b5f808273ffffffffffffffffffffffffffffffffffffffff163b119050919050565b5f604051905090565b5f80fd5b5f80fd5b5f7fffffffff0000000000000000000000000000000000000000000000000000000082169050919050565b61285f8161282b565b8114612869575f80fd5b50565b5f8135905061287a81612856565b92915050565b5f6020828403121561289557612894612823565b5b5f6128a28482850161286c565b91505092915050565b5f8115159050919050565b6128bf816128ab565b82525050565b5f6020820190506128d85f8301846128b6565b92915050565b5f81519050919050565b5f82825260208201905092915050565b5f5b838110156129155780820151818401526020810190506128fa565b5f8484015250505050565b5f601f19601f8301169050919050565b5f61293a826128de565b61294481856128e8565b93506129548185602086016128f8565b61295d81612920565b840191505092915050565b5f6020820190508181035f8301526129808184612930565b905092915050565b5f819050919050565b61299a81612988565b81146129a4575f80fd5b50565b5f813590506129b581612991565b92915050565b5f602082840312156129d0576129cf612823565b5b5f6129dd848285016129a7565b91505092915050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f612a0f826129e6565b9050919050565b612a1f81612a05565b82525050565b5f602082019050612a385f830184612a16565b92915050565b612a4781612a05565b8114612a51575f80fd5b50565b5f81359050612a6281612a3e565b92915050565b5f8060408385031215612a7e57612a7d612823565b5b5f612a8b85828601612a54565b9250506020612a9c858286016129a7565b9150509250929050565b612aaf81612988565b82525050565b5f602082019050612ac85f830184612aa6565b92915050565b5f805f60608486031215612ae557612ae4612823565b5b5f612af286828701612a54565b9350506020612b0386828701612a54565b9250506040612b14868287016129a7565b9150509250925092565b5f60208284031215612b3357612b32612823565b5b5f612b4084828501612a54565b91505092915050565b612b52816128ab565b8114612b5c575f80fd5b50565b5f81359050612b6d81612b49565b92915050565b5f8060408385031215612b8957612b88612823565b5b5f612b9685828601612a54565b9250506020612ba785828601612b5f565b9150509250929050565b5f80fd5b5f80fd5b7f4e487b71000000000000000000000000000000000000000000000000000000005f52604160045260245ffd5b612bef82612920565b810181811067ffffffffffffffff82111715612c0e57612c0d612bb9565b5b80604052505050565b5f612c2061281a565b9050612c2c8282612be6565b919050565b5f67ffffffffffffffff821115612c4b57612c4a612bb9565b5b612c5482612920565b9050602081019050919050565b828183375f83830152505050565b5f612c81612c7c84612c31565b612c17565b905082815260208101848484011115612c9d57612c9c612bb5565b5b612ca8848285612c61565b509392505050565b5f82601f830112612cc457612cc3612bb1565b5b8135612cd4848260208601612c6f565b91505092915050565b5f805f805f8060c08789031215612cf757612cf6612823565b5b5f612d0489828a016129a7565b965050602087013567ffffffffffffffff811115612d2557612d24612827565b5b612d3189828a01612cb0565b955050604087013567ffffffffffffffff811115612d5257612d51612827565b5b612d5e89828a01612cb0565b9450506060612d6f89828a016129a7565b9350506080612d8089828a016129a7565b92505060a0612d9189828a01612a54565b9150509295509295509295565b5f67ffffffffffffffff821115612db857612db7612bb9565b5b612dc182612920565b9050602081019050919050565b5f612de0612ddb84612d9e565b612c17565b905082815260208101848484011115612dfc57612dfb612bb5565b5b612e07848285612c61565b509392505050565b5f82601f830112612e2357612e22612bb1565b5b8135612e33848260208601612dce565b91505092915050565b5f805f8060808587031215612e5457612e53612823565b5b5f612e6187828801612a54565b9450506020612e7287828801612a54565b9350506040612e83878288016129a7565b925050606085013567ffffffffffffffff811115612ea457612ea3612827565b5b612eb087828801612e0f565b91505092959194509250565b5f60c082019050612ecf5f830189612aa6565b8181036020830152612ee18188612930565b90508181036040830152612ef58187612930565b9050612f046060830186612a16565b612f116080830185612aa6565b612f1e60a0830184612aa6565b979650505050505050565b5f8060408385031215612f3f57612f3e612823565b5b5f612f4c85828601612a54565b9250506020612f5d85828601612a54565b9150509250929050565b5f60a0820190508181035f830152612f7f8188612930565b90508181036020830152612f938187612930565b9050612fa26040830186612a16565b612faf6060830185612aa6565b612fbc6080830184612aa6565b9695505050505050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52602260045260245ffd5b5f600282049050600182168061300a57607f821691505b60208210810361301d5761301c612fc6565b5b50919050565b7f4552433732313a20617070726f76616c20746f2063757272656e74206f776e655f8201527f7200000000000000000000000000000000000000000000000000000000000000602082015250565b5f61307d6021836128e8565b915061308882613023565b604082019050919050565b5f6020820190508181035f8301526130aa81613071565b9050919050565b7f4552433732313a20617070726f76652063616c6c6572206973206e6f7420746f5f8201527f6b656e206f776e6572206f7220617070726f76656420666f7220616c6c000000602082015250565b5f61310b603d836128e8565b9150613116826130b1565b604082019050919050565b5f6020820190508181035f830152613138816130ff565b9050919050565b7f4552433732313a2063616c6c6572206973206e6f7420746f6b656e206f776e655f8201527f72206f7220617070726f76656400000000000000000000000000000000000000602082015250565b5f613199602d836128e8565b91506131a48261313f565b604082019050919050565b5f6020820190508181035f8301526131c68161318d565b9050919050565b7f455243373231456e756d657261626c653a206f776e657220696e646578206f755f8201527f74206f6620626f756e6473000000000000000000000000000000000000000000602082015250565b5f613227602b836128e8565b9150613232826131cd565b604082019050919050565b5f6020820190508181035f8301526132548161321b565b9050919050565b7f455243373231456e756d657261626c653a20676c6f62616c20696e646578206f5f8201527f7574206f6620626f756e64730000000000000000000000000000000000000000602082015250565b5f6132b5602c836128e8565b91506132c08261325b565b604082019050919050565b5f6020820190508181035f8301526132e2816132a9565b9050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52603260045260245ffd5b7f4552433732313a20696e76616c696420746f6b656e20494400000000000000005f82015250565b5f61334a6018836128e8565b915061335582613316565b602082019050919050565b5f6020820190508181035f8301526133778161333e565b9050919050565b7f4552433732313a2061646472657373207a65726f206973206e6f7420612076615f8201527f6c6964206f776e65720000000000000000000000000000000000000000000000602082015250565b5f6133d86029836128e8565b91506133e38261337e565b604082019050919050565b5f6020820190508181035f830152613405816133cc565b9050919050565b7f566964656f20494420616c7265616479207573656400000000000000000000005f82015250565b5f6134406015836128e8565b915061344b8261340c565b602082019050919050565b5f6020820190508181035f83015261346d81613434565b9050919050565b7f546f74616c20737570706c79206d7573742062652067726561746572207468615f8201527f6e20300000000000000000000000000000000000000000000000000000000000602082015250565b5f6134ce6023836128e8565b91506134d982613474565b604082019050919050565b5f6020820190508181035f8301526134fb816134c2565b9050919050565b7f496e76616c69642063726561746f7220616464726573730000000000000000005f82015250565b5f6135366017836128e8565b915061354182613502565b602082019050919050565b5f6020820190508181035f8301526135638161352a565b9050919050565b5f819050815f5260205f209050919050565b5f6020601f8301049050919050565b5f82821b905092915050565b5f600883026135c67fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8261358b565b6135d0868361358b565b95508019841693508086168417925050509392505050565b5f819050919050565b5f61360b61360661360184612988565b6135e8565b612988565b9050919050565b5f819050919050565b613624836135f1565b61363861363082613612565b848454613597565b825550505050565b5f90565b61364c613640565b61365781848461361b565b505050565b5b8181101561367a5761366f5f82613644565b60018101905061365d565b5050565b601f8211156136bf576136908161356a565b6136998461357c565b810160208510156136a8578190505b6136bc6136b48561357c565b83018261365c565b50505b505050565b5f82821c905092915050565b5f6136df5f19846008026136c4565b1980831691505092915050565b5f6136f783836136d0565b9150826002028217905092915050565b613710826128de565b67ffffffffffffffff81111561372957613728612bb9565b5b6137338254612ff3565b61373e82828561367e565b5f60209050601f83116001811461376f575f841561375d578287015190505b61376785826136ec565b8655506137ce565b601f19841661377d8661356a565b5f5b828110156137a45784890151825560018201915060208501945060208101905061377f565b868310156137c157848901516137bd601f8916826136d0565b8355505b6001600288020188555050505b505050505050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f61380d82612988565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff820361383f5761383e6137d6565b5b600182019050919050565b5f60808201905061385d5f830187612aa6565b818103602083015261386f8186612930565b905081810360408301526138838185612930565b90506138926060830184612aa6565b95945050505050565b5f81905092915050565b5f6138af826128de565b6138b9818561389b565b93506138c98185602086016128f8565b80840191505092915050565b5f6138e082856138a5565b91506138ec82846138a5565b91508190509392505050565b7f4f776e61626c653a206e6577206f776e657220697320746865207a65726f20615f8201527f6464726573730000000000000000000000000000000000000000000000000000602082015250565b5f6139526026836128e8565b915061395d826138f8565b604082019050919050565b5f6020820190508181035f83015261397f81613946565b9050919050565b7f4552433732313a207472616e736665722066726f6d20696e636f7272656374205f8201527f6f776e6572000000000000000000000000000000000000000000000000000000602082015250565b5f6139e06025836128e8565b91506139eb82613986565b604082019050919050565b5f6020820190508181035f830152613a0d816139d4565b9050919050565b7f4552433732313a207472616e7366657220746f20746865207a65726f206164645f8201527f7265737300000000000000000000000000000000000000000000000000000000602082015250565b5f613a6e6024836128e8565b9150613a7982613a14565b604082019050919050565b5f6020820190508181035f830152613a9b81613a62565b9050919050565b7f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e65725f82015250565b5f613ad66020836128e8565b9150613ae182613aa2565b602082019050919050565b5f6020820190508181035f830152613b0381613aca565b9050919050565b7f4552433732313a20617070726f766520746f2063616c6c6572000000000000005f82015250565b5f613b3e6019836128e8565b9150613b4982613b0a565b602082019050919050565b5f6020820190508181035f830152613b6b81613b32565b9050919050565b7f4552433732313a206d696e7420746f20746865207a65726f20616464726573735f82015250565b5f613ba66020836128e8565b9150613bb182613b72565b602082019050919050565b5f6020820190508181035f830152613bd381613b9a565b9050919050565b7f4552433732313a20746f6b656e20616c7265616479206d696e746564000000005f82015250565b5f613c0e601c836128e8565b9150613c1982613bda565b602082019050919050565b5f6020820190508181035f830152613c3b81613c02565b9050919050565b7f4552433732313a207472616e7366657220746f206e6f6e2045524337323152655f8201527f63656976657220696d706c656d656e7465720000000000000000000000000000602082015250565b5f613c9c6032836128e8565b9150613ca782613c42565b604082019050919050565b5f6020820190508181035f830152613cc981613c90565b9050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601260045260245ffd5b7f455243373231456e756d657261626c653a20636f6e73656375746976652074725f8201527f616e7366657273206e6f7420737570706f727465640000000000000000000000602082015250565b5f613d576035836128e8565b9150613d6282613cfd565b604082019050919050565b5f6020820190508181035f830152613d8481613d4b565b9050919050565b5f81519050919050565b5f82825260208201905092915050565b5f613daf82613d8b565b613db98185613d95565b9350613dc98185602086016128f8565b613dd281612920565b840191505092915050565b5f608082019050613df05f830187612a16565b613dfd6020830186612a16565b613e0a6040830185612aa6565b8181036060830152613e1c8184613da5565b905095945050505050565b5f81519050613e3581612856565b92915050565b5f60208284031215613e5057613e4f612823565b5b5f613e5d84828501613e27565b91505092915050565b5f613e7082612988565b9150613e7b83612988565b9250828203905081811115613e9357613e926137d6565b5b92915050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52603160045260245ffdfea26469706673582212201b20217e6eb500d05f95c4528ff6e49604f925cccadfd0eb87997469c5b159aa64736f6c63430008150033\n";
+
+    private static String librariesLinkedBinary;
+
+    public static final String FUNC_APPROVE = "approve";
+
+    public static final String FUNC_BALANCEOF = "balanceOf";
+
+    public static final String FUNC_GETAPPROVED = "getApproved";
+
+    public static final String FUNC_ISAPPROVEDFORALL = "isApprovedForAll";
+
+    public static final String FUNC_NAME = "name";
+
+    public static final String FUNC_OWNER = "owner";
+
+    public static final String FUNC_OWNEROF = "ownerOf";
+
+    public static final String FUNC_RENOUNCEOWNERSHIP = "renounceOwnership";
+
+    public static final String FUNC_safeTransferFrom = "safeTransferFrom";
+
+    public static final String FUNC_SETAPPROVALFORALL = "setApprovalForAll";
+
+    public static final String FUNC_SUPPORTSINTERFACE = "supportsInterface";
+
+    public static final String FUNC_SYMBOL = "symbol";
+
+    public static final String FUNC_TOKENBYINDEX = "tokenByIndex";
+
+    public static final String FUNC_TOKENOFOWNERBYINDEX = "tokenOfOwnerByIndex";
+
+    public static final String FUNC_TOKENTOVIDEO = "tokenToVideo";
+
+    public static final String FUNC_TOKENURI = "tokenURI";
+
+    public static final String FUNC_TOTALSUPPLY = "totalSupply";
+
+    public static final String FUNC_TRANSFERFROM = "transferFrom";
+
+    public static final String FUNC_TRANSFEROWNERSHIP = "transferOwnership";
+
+    public static final String FUNC_TRUSTEDOPERATOR = "trustedOperator";
+
+    public static final String FUNC_VIDEOEXISTS = "videoExists";
+
+    public static final String FUNC_VIDEOS = "videos";
+
+    public static final String FUNC_SETTRUSTEDOPERATOR = "setTrustedOperator";
+
+    public static final String FUNC_MINTVIDEONFT = "mintVideoNFT";
+
+    public static final String FUNC_GETVIDEOINFO = "getVideoInfo";
+
+    public static final String FUNC_GETVIDEOIDOFTOKEN = "getVideoIdOfToken";
+
+    public static final Event APPROVAL_EVENT =
+            new Event(
+                    "Approval",
+                    Arrays.<TypeReference<?>>asList(
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Uint256>(true) {}));
+    ;
+
+    public static final Event APPROVALFORALL_EVENT =
+            new Event(
+                    "ApprovalForAll",
+                    Arrays.<TypeReference<?>>asList(
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Bool>() {}));
+    ;
+
+    public static final Event OWNERSHIPTRANSFERRED_EVENT =
+            new Event(
+                    "OwnershipTransferred",
+                    Arrays.<TypeReference<?>>asList(
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Address>(true) {}));
+    ;
+
+    public static final Event TRANSFER_EVENT =
+            new Event(
+                    "Transfer",
+                    Arrays.<TypeReference<?>>asList(
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Uint256>(true) {}));
+    ;
+
+    public static final Event VIDEONFTMINTED_EVENT =
+            new Event(
+                    "VideoNFTMinted",
+                    Arrays.<TypeReference<?>>asList(
+                            new TypeReference<Uint256>(true) {},
+                            new TypeReference<Address>(true) {},
+                            new TypeReference<Uint256>() {},
+                            new TypeReference<Utf8String>() {},
+                            new TypeReference<Utf8String>() {},
+                            new TypeReference<Uint256>() {}));
+    ;
+
+    @Deprecated
+    protected VideoNFT(
+            String contractAddress,
+            Web3j web3j,
+            Credentials credentials,
+            BigInteger gasPrice,
+            BigInteger gasLimit) {
+        super(BINARY, contractAddress, web3j, credentials, gasPrice, gasLimit);
+    }
+
+    protected VideoNFT(
+            String contractAddress,
+            Web3j web3j,
+            Credentials credentials,
+            ContractGasProvider contractGasProvider) {
+        super(BINARY, contractAddress, web3j, credentials, contractGasProvider);
+    }
+
+    @Deprecated
+    protected VideoNFT(
+            String contractAddress,
+            Web3j web3j,
+            TransactionManager transactionManager,
+            BigInteger gasPrice,
+            BigInteger gasLimit) {
+        super(BINARY, contractAddress, web3j, transactionManager, gasPrice, gasLimit);
+    }
+
+    protected VideoNFT(
+            String contractAddress,
+            Web3j web3j,
+            TransactionManager transactionManager,
+            ContractGasProvider contractGasProvider) {
+        super(BINARY, contractAddress, web3j, transactionManager, contractGasProvider);
+    }
+
+    public static List<ApprovalEventResponse> getApprovalEvents(
+            TransactionReceipt transactionReceipt) {
+        List<Contract.EventValuesWithLog> valueList =
+                staticExtractEventParametersWithLog(APPROVAL_EVENT, transactionReceipt);
+        ArrayList<ApprovalEventResponse> responses =
+                new ArrayList<ApprovalEventResponse>(valueList.size());
+        for (Contract.EventValuesWithLog eventValues : valueList) {
+            ApprovalEventResponse typedResponse = new ApprovalEventResponse();
+            typedResponse.log = eventValues.getLog();
+            typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
+            typedResponse.approved = (String) eventValues.getIndexedValues().get(1).getValue();
+            typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
+            responses.add(typedResponse);
+        }
+        return responses;
+    }
+
+    public static ApprovalEventResponse getApprovalEventFromLog(Log log) {
+        Contract.EventValuesWithLog eventValues =
+                staticExtractEventParametersWithLog(APPROVAL_EVENT, log);
+        ApprovalEventResponse typedResponse = new ApprovalEventResponse();
+        typedResponse.log = log;
+        typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
+        typedResponse.approved = (String) eventValues.getIndexedValues().get(1).getValue();
+        typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
+        return typedResponse;
+    }
+
+    public Flowable<ApprovalEventResponse> approvalEventFlowable(EthFilter filter) {
+        return web3j.ethLogFlowable(filter).map(log -> getApprovalEventFromLog(log));
+    }
+
+    public Flowable<ApprovalEventResponse> approvalEventFlowable(
+            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
+        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
+        filter.addSingleTopic(EventEncoder.encode(APPROVAL_EVENT));
+        return approvalEventFlowable(filter);
+    }
+
+    public static List<ApprovalForAllEventResponse> getApprovalForAllEvents(
+            TransactionReceipt transactionReceipt) {
+        List<Contract.EventValuesWithLog> valueList =
+                staticExtractEventParametersWithLog(APPROVALFORALL_EVENT, transactionReceipt);
+        ArrayList<ApprovalForAllEventResponse> responses =
+                new ArrayList<ApprovalForAllEventResponse>(valueList.size());
+        for (Contract.EventValuesWithLog eventValues : valueList) {
+            ApprovalForAllEventResponse typedResponse = new ApprovalForAllEventResponse();
+            typedResponse.log = eventValues.getLog();
+            typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
+            typedResponse.operator = (String) eventValues.getIndexedValues().get(1).getValue();
+            typedResponse.approved = (Boolean) eventValues.getNonIndexedValues().get(0).getValue();
+            responses.add(typedResponse);
+        }
+        return responses;
+    }
+
+    public static ApprovalForAllEventResponse getApprovalForAllEventFromLog(Log log) {
+        Contract.EventValuesWithLog eventValues =
+                staticExtractEventParametersWithLog(APPROVALFORALL_EVENT, log);
+        ApprovalForAllEventResponse typedResponse = new ApprovalForAllEventResponse();
+        typedResponse.log = log;
+        typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
+        typedResponse.operator = (String) eventValues.getIndexedValues().get(1).getValue();
+        typedResponse.approved = (Boolean) eventValues.getNonIndexedValues().get(0).getValue();
+        return typedResponse;
+    }
+
+    public Flowable<ApprovalForAllEventResponse> approvalForAllEventFlowable(EthFilter filter) {
+        return web3j.ethLogFlowable(filter).map(log -> getApprovalForAllEventFromLog(log));
+    }
+
+    public Flowable<ApprovalForAllEventResponse> approvalForAllEventFlowable(
+            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
+        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
+        filter.addSingleTopic(EventEncoder.encode(APPROVALFORALL_EVENT));
+        return approvalForAllEventFlowable(filter);
+    }
+
+    public static List<OwnershipTransferredEventResponse> getOwnershipTransferredEvents(
+            TransactionReceipt transactionReceipt) {
+        List<Contract.EventValuesWithLog> valueList =
+                staticExtractEventParametersWithLog(OWNERSHIPTRANSFERRED_EVENT, transactionReceipt);
+        ArrayList<OwnershipTransferredEventResponse> responses =
+                new ArrayList<OwnershipTransferredEventResponse>(valueList.size());
+        for (Contract.EventValuesWithLog eventValues : valueList) {
+            OwnershipTransferredEventResponse typedResponse =
+                    new OwnershipTransferredEventResponse();
+            typedResponse.log = eventValues.getLog();
+            typedResponse.previousOwner = (String) eventValues.getIndexedValues().get(0).getValue();
+            typedResponse.newOwner = (String) eventValues.getIndexedValues().get(1).getValue();
+            responses.add(typedResponse);
+        }
+        return responses;
+    }
+
+    public static OwnershipTransferredEventResponse getOwnershipTransferredEventFromLog(Log log) {
+        Contract.EventValuesWithLog eventValues =
+                staticExtractEventParametersWithLog(OWNERSHIPTRANSFERRED_EVENT, log);
+        OwnershipTransferredEventResponse typedResponse = new OwnershipTransferredEventResponse();
+        typedResponse.log = log;
+        typedResponse.previousOwner = (String) eventValues.getIndexedValues().get(0).getValue();
+        typedResponse.newOwner = (String) eventValues.getIndexedValues().get(1).getValue();
+        return typedResponse;
+    }
+
+    public Flowable<OwnershipTransferredEventResponse> ownershipTransferredEventFlowable(
+            EthFilter filter) {
+        return web3j.ethLogFlowable(filter).map(log -> getOwnershipTransferredEventFromLog(log));
+    }
+
+    public Flowable<OwnershipTransferredEventResponse> ownershipTransferredEventFlowable(
+            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
+        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
+        filter.addSingleTopic(EventEncoder.encode(OWNERSHIPTRANSFERRED_EVENT));
+        return ownershipTransferredEventFlowable(filter);
+    }
+
+    public static List<TransferEventResponse> getTransferEvents(
+            TransactionReceipt transactionReceipt) {
+        List<Contract.EventValuesWithLog> valueList =
+                staticExtractEventParametersWithLog(TRANSFER_EVENT, transactionReceipt);
+        ArrayList<TransferEventResponse> responses =
+                new ArrayList<TransferEventResponse>(valueList.size());
+        for (Contract.EventValuesWithLog eventValues : valueList) {
+            TransferEventResponse typedResponse = new TransferEventResponse();
+            typedResponse.log = eventValues.getLog();
+            typedResponse.from = (String) eventValues.getIndexedValues().get(0).getValue();
+            typedResponse.to = (String) eventValues.getIndexedValues().get(1).getValue();
+            typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
+            responses.add(typedResponse);
+        }
+        return responses;
+    }
+
+    public static TransferEventResponse getTransferEventFromLog(Log log) {
+        Contract.EventValuesWithLog eventValues =
+                staticExtractEventParametersWithLog(TRANSFER_EVENT, log);
+        TransferEventResponse typedResponse = new TransferEventResponse();
+        typedResponse.log = log;
+        typedResponse.from = (String) eventValues.getIndexedValues().get(0).getValue();
+        typedResponse.to = (String) eventValues.getIndexedValues().get(1).getValue();
+        typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
+        return typedResponse;
+    }
+
+    public Flowable<TransferEventResponse> transferEventFlowable(EthFilter filter) {
+        return web3j.ethLogFlowable(filter).map(log -> getTransferEventFromLog(log));
+    }
+
+    public Flowable<TransferEventResponse> transferEventFlowable(
+            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
+        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
+        filter.addSingleTopic(EventEncoder.encode(TRANSFER_EVENT));
+        return transferEventFlowable(filter);
+    }
+
+    public static List<VideoNFTMintedEventResponse> getVideoNFTMintedEvents(
+            TransactionReceipt transactionReceipt) {
+        List<Contract.EventValuesWithLog> valueList =
+                staticExtractEventParametersWithLog(VIDEONFTMINTED_EVENT, transactionReceipt);
+        ArrayList<VideoNFTMintedEventResponse> responses =
+                new ArrayList<VideoNFTMintedEventResponse>(valueList.size());
+        for (Contract.EventValuesWithLog eventValues : valueList) {
+            VideoNFTMintedEventResponse typedResponse = new VideoNFTMintedEventResponse();
+            typedResponse.log = eventValues.getLog();
+            typedResponse.videoId = (BigInteger) eventValues.getIndexedValues().get(0).getValue();
+            typedResponse.creatorAddress =
+                    (String) eventValues.getIndexedValues().get(1).getValue();
+            typedResponse.totalSupply =
+                    (BigInteger) eventValues.getNonIndexedValues().get(0).getValue();
+            typedResponse.name = (String) eventValues.getNonIndexedValues().get(1).getValue();
+            typedResponse.symbol = (String) eventValues.getNonIndexedValues().get(2).getValue();
+            typedResponse.price = (BigInteger) eventValues.getNonIndexedValues().get(3).getValue();
+            responses.add(typedResponse);
+        }
+        return responses;
+    }
+
+    public static VideoNFTMintedEventResponse getVideoNFTMintedEventFromLog(Log log) {
+        Contract.EventValuesWithLog eventValues =
+                staticExtractEventParametersWithLog(VIDEONFTMINTED_EVENT, log);
+        VideoNFTMintedEventResponse typedResponse = new VideoNFTMintedEventResponse();
+        typedResponse.log = log;
+        typedResponse.videoId = (BigInteger) eventValues.getIndexedValues().get(0).getValue();
+        typedResponse.creatorAddress = (String) eventValues.getIndexedValues().get(1).getValue();
+        typedResponse.totalSupply =
+                (BigInteger) eventValues.getNonIndexedValues().get(0).getValue();
+        typedResponse.name = (String) eventValues.getNonIndexedValues().get(1).getValue();
+        typedResponse.symbol = (String) eventValues.getNonIndexedValues().get(2).getValue();
+        typedResponse.price = (BigInteger) eventValues.getNonIndexedValues().get(3).getValue();
+        return typedResponse;
+    }
+
+    public Flowable<VideoNFTMintedEventResponse> videoNFTMintedEventFlowable(EthFilter filter) {
+        return web3j.ethLogFlowable(filter).map(log -> getVideoNFTMintedEventFromLog(log));
+    }
+
+    public Flowable<VideoNFTMintedEventResponse> videoNFTMintedEventFlowable(
+            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
+        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
+        filter.addSingleTopic(EventEncoder.encode(VIDEONFTMINTED_EVENT));
+        return videoNFTMintedEventFlowable(filter);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> approve(String to, BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_APPROVE,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, to),
+                                new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<BigInteger> balanceOf(String owner) {
+        final Function function =
+                new Function(
+                        FUNC_BALANCEOF,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.Address(160, owner)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
+        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
+    }
+
+    public RemoteFunctionCall<String> getApproved(BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_GETAPPROVED,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<Boolean> isApprovedForAll(String owner, String operator) {
+        final Function function =
+                new Function(
+                        FUNC_ISAPPROVEDFORALL,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, owner),
+                                new org.web3j.abi.datatypes.Address(160, operator)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Bool>() {}));
+        return executeRemoteCallSingleValueReturn(function, Boolean.class);
+    }
+
+    public RemoteFunctionCall<String> name() {
+        final Function function =
+                new Function(
+                        FUNC_NAME,
+                        Arrays.<Type>asList(),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Utf8String>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<String> owner() {
+        final Function function =
+                new Function(
+                        FUNC_OWNER,
+                        Arrays.<Type>asList(),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<String> ownerOf(BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_OWNEROF,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> renounceOwnership() {
+        final Function function =
+                new Function(
+                        FUNC_RENOUNCEOWNERSHIP,
+                        Arrays.<Type>asList(),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> safeTransferFrom(
+            String from, String to, BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_safeTransferFrom,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, from),
+                                new org.web3j.abi.datatypes.Address(160, to),
+                                new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> safeTransferFrom(
+            String from, String to, BigInteger tokenId, byte[] data) {
+        final Function function =
+                new Function(
+                        FUNC_safeTransferFrom,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, from),
+                                new org.web3j.abi.datatypes.Address(160, to),
+                                new org.web3j.abi.datatypes.generated.Uint256(tokenId),
+                                new org.web3j.abi.datatypes.DynamicBytes(data)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> setApprovalForAll(
+            String operator, Boolean approved) {
+        final Function function =
+                new Function(
+                        FUNC_SETAPPROVALFORALL,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, operator),
+                                new org.web3j.abi.datatypes.Bool(approved)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<Boolean> supportsInterface(byte[] interfaceId) {
+        final Function function =
+                new Function(
+                        FUNC_SUPPORTSINTERFACE,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.generated.Bytes4(interfaceId)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Bool>() {}));
+        return executeRemoteCallSingleValueReturn(function, Boolean.class);
+    }
+
+    public RemoteFunctionCall<String> symbol() {
+        final Function function =
+                new Function(
+                        FUNC_SYMBOL,
+                        Arrays.<Type>asList(),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Utf8String>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<BigInteger> tokenByIndex(BigInteger index) {
+        final Function function =
+                new Function(
+                        FUNC_TOKENBYINDEX,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(index)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
+        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
+    }
+
+    public RemoteFunctionCall<BigInteger> tokenOfOwnerByIndex(String owner, BigInteger index) {
+        final Function function =
+                new Function(
+                        FUNC_TOKENOFOWNERBYINDEX,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, owner),
+                                new org.web3j.abi.datatypes.generated.Uint256(index)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
+        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
+    }
+
+    public RemoteFunctionCall<BigInteger> tokenToVideo(BigInteger param0) {
+        final Function function =
+                new Function(
+                        FUNC_TOKENTOVIDEO,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(param0)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
+        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
+    }
+
+    public RemoteFunctionCall<String> tokenURI(BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_TOKENURI,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Utf8String>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<BigInteger> totalSupply() {
+        final Function function =
+                new Function(
+                        FUNC_TOTALSUPPLY,
+                        Arrays.<Type>asList(),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
+        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> transferFrom(
+            String from, String to, BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_TRANSFERFROM,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, from),
+                                new org.web3j.abi.datatypes.Address(160, to),
+                                new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> transferOwnership(String newOwner) {
+        final Function function =
+                new Function(
+                        FUNC_TRANSFEROWNERSHIP,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.Address(160, newOwner)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<String> trustedOperator() {
+        final Function function =
+                new Function(
+                        FUNC_TRUSTEDOPERATOR,
+                        Arrays.<Type>asList(),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
+        return executeRemoteCallSingleValueReturn(function, String.class);
+    }
+
+    public RemoteFunctionCall<Boolean> videoExists(BigInteger param0) {
+        final Function function =
+                new Function(
+                        FUNC_VIDEOEXISTS,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(param0)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Bool>() {}));
+        return executeRemoteCallSingleValueReturn(function, Boolean.class);
+    }
+
+    public RemoteFunctionCall<Tuple6<BigInteger, String, String, String, BigInteger, BigInteger>>
+            videos(BigInteger param0) {
+        final Function function =
+                new Function(
+                        FUNC_VIDEOS,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(param0)),
+                        Arrays.<TypeReference<?>>asList(
+                                new TypeReference<Uint256>() {},
+                                new TypeReference<Utf8String>() {},
+                                new TypeReference<Utf8String>() {},
+                                new TypeReference<Address>() {},
+                                new TypeReference<Uint256>() {},
+                                new TypeReference<Uint256>() {}));
+        return new RemoteFunctionCall<
+                Tuple6<BigInteger, String, String, String, BigInteger, BigInteger>>(
+                function,
+                new Callable<Tuple6<BigInteger, String, String, String, BigInteger, BigInteger>>() {
+                    @Override
+                    public Tuple6<BigInteger, String, String, String, BigInteger, BigInteger> call()
+                            throws Exception {
+                        List<Type> results = executeCallMultipleValueReturn(function);
+                        return new Tuple6<
+                                BigInteger, String, String, String, BigInteger, BigInteger>(
+                                (BigInteger) results.get(0).getValue(),
+                                (String) results.get(1).getValue(),
+                                (String) results.get(2).getValue(),
+                                (String) results.get(3).getValue(),
+                                (BigInteger) results.get(4).getValue(),
+                                (BigInteger) results.get(5).getValue());
+                    }
+                });
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> setTrustedOperator(String _operator) {
+        final Function function =
+                new Function(
+                        FUNC_SETTRUSTEDOPERATOR,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.Address(160, _operator)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<TransactionReceipt> mintVideoNFT(
+            BigInteger videoId,
+            String nftName,
+            String nftSymbol,
+            BigInteger totalSupply,
+            BigInteger price,
+            String creatorAddress) {
+        final Function function =
+                new Function(
+                        FUNC_MINTVIDEONFT,
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.generated.Uint256(videoId),
+                                new org.web3j.abi.datatypes.Utf8String(nftName),
+                                new org.web3j.abi.datatypes.Utf8String(nftSymbol),
+                                new org.web3j.abi.datatypes.generated.Uint256(totalSupply),
+                                new org.web3j.abi.datatypes.generated.Uint256(price),
+                                new org.web3j.abi.datatypes.Address(160, creatorAddress)),
+                        Collections.<TypeReference<?>>emptyList());
+        return executeRemoteCallTransaction(function);
+    }
+
+    public RemoteFunctionCall<Tuple5<String, String, String, BigInteger, BigInteger>> getVideoInfo(
+            BigInteger videoId) {
+        final Function function =
+                new Function(
+                        FUNC_GETVIDEOINFO,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(videoId)),
+                        Arrays.<TypeReference<?>>asList(
+                                new TypeReference<Utf8String>() {},
+                                new TypeReference<Utf8String>() {},
+                                new TypeReference<Address>() {},
+                                new TypeReference<Uint256>() {},
+                                new TypeReference<Uint256>() {}));
+        return new RemoteFunctionCall<Tuple5<String, String, String, BigInteger, BigInteger>>(
+                function,
+                new Callable<Tuple5<String, String, String, BigInteger, BigInteger>>() {
+                    @Override
+                    public Tuple5<String, String, String, BigInteger, BigInteger> call()
+                            throws Exception {
+                        List<Type> results = executeCallMultipleValueReturn(function);
+                        return new Tuple5<String, String, String, BigInteger, BigInteger>(
+                                (String) results.get(0).getValue(),
+                                (String) results.get(1).getValue(),
+                                (String) results.get(2).getValue(),
+                                (BigInteger) results.get(3).getValue(),
+                                (BigInteger) results.get(4).getValue());
+                    }
+                });
+    }
+
+    public RemoteFunctionCall<BigInteger> getVideoIdOfToken(BigInteger tokenId) {
+        final Function function =
+                new Function(
+                        FUNC_GETVIDEOIDOFTOKEN,
+                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
+                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
+        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
+    }
+
+    @Deprecated
+    public static VideoNFT load(
+            String contractAddress,
+            Web3j web3j,
+            Credentials credentials,
+            BigInteger gasPrice,
+            BigInteger gasLimit) {
+        return new VideoNFT(contractAddress, web3j, credentials, gasPrice, gasLimit);
+    }
+
+    @Deprecated
+    public static VideoNFT load(
+            String contractAddress,
+            Web3j web3j,
+            TransactionManager transactionManager,
+            BigInteger gasPrice,
+            BigInteger gasLimit) {
+        return new VideoNFT(contractAddress, web3j, transactionManager, gasPrice, gasLimit);
+    }
+
+    public static VideoNFT load(
+            String contractAddress,
+            Web3j web3j,
+            Credentials credentials,
+            ContractGasProvider contractGasProvider) {
+        return new VideoNFT(contractAddress, web3j, credentials, contractGasProvider);
+    }
+
+    public static VideoNFT load(
+            String contractAddress,
+            Web3j web3j,
+            TransactionManager transactionManager,
+            ContractGasProvider contractGasProvider) {
+        return new VideoNFT(contractAddress, web3j, transactionManager, contractGasProvider);
+    }
+
+    public static RemoteCall<VideoNFT> deploy(
+            Web3j web3j,
+            Credentials credentials,
+            ContractGasProvider contractGasProvider,
+            String name,
+            String symbol,
+            String _trustedOperator) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Utf8String(name),
+                                new org.web3j.abi.datatypes.Utf8String(symbol),
+                                new org.web3j.abi.datatypes.Address(160, _trustedOperator)));
+        return deployRemoteCall(
+                VideoNFT.class,
+                web3j,
+                credentials,
+                contractGasProvider,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    public static RemoteCall<VideoNFT> deploy(
+            Web3j web3j,
+            TransactionManager transactionManager,
+            ContractGasProvider contractGasProvider,
+            String name,
+            String symbol,
+            String _trustedOperator) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Utf8String(name),
+                                new org.web3j.abi.datatypes.Utf8String(symbol),
+                                new org.web3j.abi.datatypes.Address(160, _trustedOperator)));
+        return deployRemoteCall(
+                VideoNFT.class,
+                web3j,
+                transactionManager,
+                contractGasProvider,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    @Deprecated
+    public static RemoteCall<VideoNFT> deploy(
+            Web3j web3j,
+            Credentials credentials,
+            BigInteger gasPrice,
+            BigInteger gasLimit,
+            String name,
+            String symbol,
+            String _trustedOperator) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Utf8String(name),
+                                new org.web3j.abi.datatypes.Utf8String(symbol),
+                                new org.web3j.abi.datatypes.Address(160, _trustedOperator)));
+        return deployRemoteCall(
+                VideoNFT.class,
+                web3j,
+                credentials,
+                gasPrice,
+                gasLimit,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    @Deprecated
+    public static RemoteCall<VideoNFT> deploy(
+            Web3j web3j,
+            TransactionManager transactionManager,
+            BigInteger gasPrice,
+            BigInteger gasLimit,
+            String name,
+            String symbol,
+            String _trustedOperator) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Utf8String(name),
+                                new org.web3j.abi.datatypes.Utf8String(symbol),
+                                new org.web3j.abi.datatypes.Address(160, _trustedOperator)));
+        return deployRemoteCall(
+                VideoNFT.class,
+                web3j,
+                transactionManager,
+                gasPrice,
+                gasLimit,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    //    public static void linkLibraries(List<Contract.LinkReference> references) {
+    //        librariesLinkedBinary = linkBinaryWithReferences(BINARY, references);
+    //    }
+
+    private static String getDeploymentBinary() {
+        return BINARY;
+    }
+
+    public static class ApprovalEventResponse extends BaseEventResponse {
+        public String owner;
+
+        public String approved;
+
+        public BigInteger tokenId;
+    }
+
+    public static class ApprovalForAllEventResponse extends BaseEventResponse {
+        public String owner;
+
+        public String operator;
+
+        public Boolean approved;
+    }
+
+    public static class OwnershipTransferredEventResponse extends BaseEventResponse {
+        public String previousOwner;
+
+        public String newOwner;
+    }
+
+    public static class TransferEventResponse extends BaseEventResponse {
+        public String from;
+
+        public String to;
+
+        public BigInteger tokenId;
+    }
+
+    public static class VideoNFTMintedEventResponse extends BaseEventResponse {
+        public BigInteger videoId;
+
+        public String creatorAddress;
+
+        public BigInteger totalSupply;
+
+        public String name;
+
+        public String symbol;
+
+        public BigInteger price;
+    }
+}
