+++ src/main/java/TokenUs/TokenUs_BE/web3/contract/VideoNFTMarketplace.java
@@ -9,6 +9,7 @@ import java.util.concurrent.Callable;
 
 import io.reactivex.Flowable;
 import org.web3j.abi.EventEncoder;
+import org.web3j.abi.FunctionEncoder;
 import org.web3j.abi.TypeReference;
 import org.web3j.abi.datatypes.Address;
 import org.web3j.abi.datatypes.Bool;
@@ -20,6 +21,7 @@ import org.web3j.abi.datatypes.generated.Uint256;
 import org.web3j.crypto.Credentials;
 import org.web3j.protocol.Web3j;
 import org.web3j.protocol.core.DefaultBlockParameter;
+import org.web3j.protocol.core.RemoteCall;
 import org.web3j.protocol.core.RemoteFunctionCall;
 import org.web3j.protocol.core.methods.request.EthFilter;
 import org.web3j.protocol.core.methods.response.BaseEventResponse;
@@ -38,13 +40,16 @@ import org.web3j.tx.gas.ContractGasProvider;
  *
  * <p>Please use the <a href="https://docs.web3j.io/command_line.html">web3j command line tools</a>,
  * or the org.web3j.codegen.SolidityFunctionWrapperGenerator in the <a
- * href="https://github.com/hyperledger-web3j/web3j/tree/main/codegen">codegen module</a> to update.
+ * href="https://github.com/LFDT-web3j/web3j/tree/main/codegen">codegen module</a> to update.
  *
- * <p>Generated with web3j version 1.6.3.
+ * <p>Generated with web3j version 1.7.0.
  */
 @SuppressWarnings("rawtypes")
-public class VideoNftMarketplace_ABI extends Contract {
-    public static final String BINARY = "Bin file was not provided";
+public class VideoNFTMarketplace extends Contract {
+    public static final String BINARY =
+            "0x608060405234801562000010575f80fd5b5060405162002369380380620023698339818101604052810190620000369190620001ca565b620000566200004a6200009d60201b60201c565b620000a460201b60201c565b8060015f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555050620001fa565b5f33905090565b5f805f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050815f806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b5f80fd5b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f620001948262000169565b9050919050565b620001a68162000188565b8114620001b1575f80fd5b50565b5f81519050620001c4816200019b565b92915050565b5f60208284031215620001e257620001e162000165565b5b5f620001f184828501620001b4565b91505092915050565b61216180620002085f395ff3fe608060405260043610610090575f3560e01c806394383f141161005857806394383f1414610144578063cb917b9c1461016c578063de74e57b14610194578063f2fde38b146101d3578063ff721aba146101fb57610090565b8063012386f014610094578063150bde03146100c0578063242cae9f146100dc578063715018a6146101045780638da5cb5b1461011a575b5f80fd5b34801561009f575f80fd5b506100a8610237565b6040516100b793929190611908565b60405180910390f35b6100da60048036038101906100d59190611980565b610680565b005b3480156100e7575f80fd5b5061010260048036038101906100fd91906119d5565b610bef565b005b34801561010f575f80fd5b50610118610c9e565b005b348015610125575f80fd5b5061012e610cb1565b60405161013b9190611a0f565b60405180910390f35b34801561014f575f80fd5b5061016a60048036038101906101659190611a28565b610cd8565b005b348015610177575f80fd5b50610192600480360381019061018d9190611980565b6110c7565b005b34801561019f575f80fd5b506101ba60048036038101906101b59190611980565b611525565b6040516101ca9493929190611a8f565b60405180910390f35b3480156101de575f80fd5b506101f960048036038101906101f491906119d5565b61157c565b005b348015610206575f80fd5b50610221600480360381019061021c91906119d5565b6115fe565b60405161022e9190611ad2565b60405180910390f35b60608060605f60015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166318160ddd6040518163ffffffff1660e01b8152600401602060405180830381865afa1580156102a7573d5f803e3d5ffd5b505050506040513d601f19601f820116820180604052508101906102cb9190611aff565b90505f805b828110156103bd575f60015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16634f6ccce7836040518263ffffffff1660e01b81526004016103339190611b2a565b602060405180830381865afa15801561034e573d5f803e3d5ffd5b505050506040513d601f19601f820116820180604052508101906103729190611aff565b905060025f8281526020019081526020015f206003015f9054906101000a900460ff16156103a95782806103a590611b70565b9350505b5080806103b590611b70565b9150506102d0565b508067ffffffffffffffff8111156103d8576103d7611bb7565b5b6040519080825280602002602001820160405280156104065781602001602082028036833780820191505090505b5094508067ffffffffffffffff81111561042357610422611bb7565b5b6040519080825280602002602001820160405280156104515781602001602082028036833780820191505090505b5093508067ffffffffffffffff81111561046e5761046d611bb7565b5b60405190808252806020026020018201604052801561049c5781602001602082028036833780820191505090505b5092505f805b83811015610677575f60015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16634f6ccce7836040518263ffffffff1660e01b81526004016105059190611b2a565b602060405180830381865afa158015610520573d5f803e3d5ffd5b505050506040513d601f19601f820116820180604052508101906105449190611aff565b905060025f8281526020019081526020015f206003015f9054906101000a900460ff16156106635760025f8281526020019081526020015f205f015488848151811061059357610592611be4565b5b60200260200101818152505060025f8281526020019081526020015f206001015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff168784815181106105e6576105e5611be4565b5b602002602001019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff168152505060025f8281526020019081526020015f206002015486848151811061064857610647611be4565b5b602002602001018181525050828061065f90611b70565b9350505b50808061066f90611b70565b9150506104a2565b50505050909192565b5f60025f8381526020019081526020015f206040518060800160405290815f8201548152602001600182015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200160028201548152602001600382015f9054906101000a900460ff16151515158152505090508060600151610765576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161075c90611c6b565b60405180910390fd5b80604001513410156107ac576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107a390611cd3565b60405180910390fd5b5f60015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636352211e846040518263ffffffff1660e01b81526004016108079190611b2a565b602060405180830381865afa158015610822573d5f803e3d5ffd5b505050506040513d601f19601f820116820180604052508101906108469190611d05565b9050816020015173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff16146108ba576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016108b190611da0565b60405180910390fd5b3073ffffffffffffffffffffffffffffffffffffffff1660015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663081812fc856040518263ffffffff1660e01b815260040161092b9190611b2a565b602060405180830381865afa158015610946573d5f803e3d5ffd5b505050506040513d601f19601f8201168201806040525081019061096a9190611d05565b73ffffffffffffffffffffffffffffffffffffffff161480610a23575060015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e985e9c582306040518363ffffffff1660e01b81526004016109e3929190611dbe565b602060405180830381865afa1580156109fe573d5f803e3d5ffd5b505050506040513d601f19601f82011682018060405250810190610a229190611e0f565b5b610a62576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610a5990611eaa565b60405180910390fd5b816020015173ffffffffffffffffffffffffffffffffffffffff166108fc836040015190811502906040515f60405180830381858888f19350505050158015610aad573d5f803e3d5ffd5b5060015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166342842e0e836020015133866040518463ffffffff1660e01b8152600401610b1093929190611ec8565b5f604051808303815f87803b158015610b27575f80fd5b505af1158015610b39573d5f803e3d5ffd5b5050505060025f8481526020019081526020015f205f8082015f9055600182015f6101000a81549073ffffffffffffffffffffffffffffffffffffffff0219169055600282015f9055600382015f6101000a81549060ff021916905550503373ffffffffffffffffffffffffffffffffffffffff16837fad546cba5af920bc2c64678bbd69f3d8965ac26d1b82f7ebf6f098aaa83352828460400151604051610be29190611b2a565b60405180910390a3505050565b610bf761161b565b600160035f8373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f6101000a81548160ff0219169083151502179055508073ffffffffffffffffffffffffffffffffffffffff167f49d13b3f4446630d8e5b7b18fad187497eb172f89d71c17ecc0e3ff12a7767136001604051610c939190611ad2565b60405180910390a250565b610ca661161b565b610caf5f611699565b565b5f805f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b5f60015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636352211e846040518263ffffffff1660e01b8152600401610d339190611b2a565b602060405180830381865afa158015610d4e573d5f803e3d5ffd5b505050506040513d601f19601f82011682018060405250810190610d729190611d05565b90505f8173ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161490505f60035f3373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f9054906101000a900460ff168015610f6257503073ffffffffffffffffffffffffffffffffffffffff1660015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663081812fc876040518263ffffffff1660e01b8152600401610e699190611b2a565b602060405180830381865afa158015610e84573d5f803e3d5ffd5b505050506040513d601f19601f82011682018060405250810190610ea89190611d05565b73ffffffffffffffffffffffffffffffffffffffff161480610f61575060015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e985e9c584306040518363ffffffff1660e01b8152600401610f21929190611dbe565b602060405180830381865afa158015610f3c573d5f803e3d5ffd5b505050506040513d601f19601f82011682018060405250810190610f609190611e0f565b5b5b90508180610f6d5750805b610fac576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610fa390611f47565b60405180910390fd5b60405180608001604052808681526020018473ffffffffffffffffffffffffffffffffffffffff1681526020018581526020016001151581525060025f8781526020019081526020015f205f820151815f01556020820151816001015f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550604082015181600201556060820151816003015f6101000a81548160ff0219169083151502179055509050508273ffffffffffffffffffffffffffffffffffffffff16857f5f9c03de076f5063cd3ebcfe8ef7aa8af378e4b35a83c77ba9777e97083879a2866040516110b89190611b2a565b60405180910390a35050505050565b5f60025f8381526020019081526020015f206040518060800160405290815f8201548152602001600182015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200160028201548152602001600382015f9054906101000a900460ff161515151581525050905080606001516111ac576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016111a390611faf565b60405180910390fd5b5f60015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16636352211e846040518263ffffffff1660e01b81526004016112079190611b2a565b602060405180830381865afa158015611222573d5f803e3d5ffd5b505050506040513d601f19601f820116820180604052508101906112469190611d05565b90505f8173ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161490505f60035f3373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020015f205f9054906101000a900460ff16801561143657503073ffffffffffffffffffffffffffffffffffffffff1660015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663081812fc876040518263ffffffff1660e01b815260040161133d9190611b2a565b602060405180830381865afa158015611358573d5f803e3d5ffd5b505050506040513d601f19601f8201168201806040525081019061137c9190611d05565b73ffffffffffffffffffffffffffffffffffffffff161480611435575060015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663e985e9c584306040518363ffffffff1660e01b81526004016113f5929190611dbe565b602060405180830381865afa158015611410573d5f803e3d5ffd5b505050506040513d601f19601f820116820180604052508101906114349190611e0f565b5b5b905081806114415750805b611480576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161147790612017565b60405180910390fd5b60025f8681526020019081526020015f205f8082015f9055600182015f6101000a81549073ffffffffffffffffffffffffffffffffffffffff0219169055600282015f9055600382015f6101000a81549060ff021916905550503373ffffffffffffffffffffffffffffffffffffffff16857fef1499b350e05776e60cd2ffcac001569a28e9d574e1f3adc1b3ab264f8ccffd60405160405180910390a35050505050565b6002602052805f5260405f205f91509050805f015490806001015f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1690806002015490806003015f9054906101000a900460ff16905084565b61158461161b565b5f73ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff16036115f2576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016115e9906120a5565b60405180910390fd5b6115fb81611699565b50565b6003602052805f5260405f205f915054906101000a900460ff1681565b61162361175a565b73ffffffffffffffffffffffffffffffffffffffff16611641610cb1565b73ffffffffffffffffffffffffffffffffffffffff1614611697576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161168e9061210d565b60405180910390fd5b565b5f805f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050815f806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b5f33905090565b5f81519050919050565b5f82825260208201905092915050565b5f819050602082019050919050565b5f819050919050565b61179c8161178a565b82525050565b5f6117ad8383611793565b60208301905092915050565b5f602082019050919050565b5f6117cf82611761565b6117d9818561176b565b93506117e48361177b565b805f5b838110156118145781516117fb88826117a2565b9750611806836117b9565b9250506001810190506117e7565b5085935050505092915050565b5f81519050919050565b5f82825260208201905092915050565b5f819050602082019050919050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f6118738261184a565b9050919050565b61188381611869565b82525050565b5f611894838361187a565b60208301905092915050565b5f602082019050919050565b5f6118b682611821565b6118c0818561182b565b93506118cb8361183b565b805f5b838110156118fb5781516118e28882611889565b97506118ed836118a0565b9250506001810190506118ce565b5085935050505092915050565b5f6060820190508181035f83015261192081866117c5565b9050818103602083015261193481856118ac565b9050818103604083015261194881846117c5565b9050949350505050565b5f80fd5b61195f8161178a565b8114611969575f80fd5b50565b5f8135905061197a81611956565b92915050565b5f6020828403121561199557611994611952565b5b5f6119a28482850161196c565b91505092915050565b6119b481611869565b81146119be575f80fd5b50565b5f813590506119cf816119ab565b92915050565b5f602082840312156119ea576119e9611952565b5b5f6119f7848285016119c1565b91505092915050565b611a0981611869565b82525050565b5f602082019050611a225f830184611a00565b92915050565b5f8060408385031215611a3e57611a3d611952565b5b5f611a4b8582860161196c565b9250506020611a5c8582860161196c565b9150509250929050565b611a6f8161178a565b82525050565b5f8115159050919050565b611a8981611a75565b82525050565b5f608082019050611aa25f830187611a66565b611aaf6020830186611a00565b611abc6040830185611a66565b611ac96060830184611a80565b95945050505050565b5f602082019050611ae55f830184611a80565b92915050565b5f81519050611af981611956565b92915050565b5f60208284031215611b1457611b13611952565b5b5f611b2184828501611aeb565b91505092915050565b5f602082019050611b3d5f830184611a66565b92915050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f611b7a8261178a565b91507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8203611bac57611bab611b43565b5b600182019050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52604160045260245ffd5b7f4e487b71000000000000000000000000000000000000000000000000000000005f52603260045260245ffd5b5f82825260208201905092915050565b7f4e4654206973206e6f74206c697374656420666f722073616c650000000000005f82015250565b5f611c55601a83611c11565b9150611c6082611c21565b602082019050919050565b5f6020820190508181035f830152611c8281611c49565b9050919050565b7f496e73756666696369656e74207061796d656e740000000000000000000000005f82015250565b5f611cbd601483611c11565b9150611cc882611c89565b602082019050919050565b5f6020820190508181035f830152611cea81611cb1565b9050919050565b5f81519050611cff816119ab565b92915050565b5f60208284031215611d1a57611d19611952565b5b5f611d2784828501611cf1565b91505092915050565b7f4c697374696e672073656c6c6572206973206e6f742063757272656e74206f775f8201527f6e65720000000000000000000000000000000000000000000000000000000000602082015250565b5f611d8a602383611c11565b9150611d9582611d30565b604082019050919050565b5f6020820190508181035f830152611db781611d7e565b9050919050565b5f604082019050611dd15f830185611a00565b611dde6020830184611a00565b9392505050565b611dee81611a75565b8114611df8575f80fd5b50565b5f81519050611e0981611de5565b92915050565b5f60208284031215611e2457611e23611952565b5b5f611e3184828501611dfb565b91505092915050565b7f4d61726b6574706c616365206973206e6f7420617070726f76656420746f20745f8201527f72616e736665722074686973204e465400000000000000000000000000000000602082015250565b5f611e94603083611c11565b9150611e9f82611e3a565b604082019050919050565b5f6020820190508181035f830152611ec181611e88565b9050919050565b5f606082019050611edb5f830186611a00565b611ee86020830185611a00565b611ef56040830184611a66565b949350505050565b7f4e6f7420617574686f72697a656420746f206c6973742074686973204e4654005f82015250565b5f611f31601f83611c11565b9150611f3c82611efd565b602082019050919050565b5f6020820190508181035f830152611f5e81611f25565b9050919050565b7f4e4654206973206e6f74206c69737465640000000000000000000000000000005f82015250565b5f611f99601183611c11565b9150611fa482611f65565b602082019050919050565b5f6020820190508181035f830152611fc681611f8d565b9050919050565b7f4e6f7420617574686f72697a656420746f2064656c69737400000000000000005f82015250565b5f612001601883611c11565b915061200c82611fcd565b602082019050919050565b5f6020820190508181035f83015261202e81611ff5565b9050919050565b7f4f776e61626c653a206e6577206f776e657220697320746865207a65726f20615f8201527f6464726573730000000000000000000000000000000000000000000000000000602082015250565b5f61208f602683611c11565b915061209a82612035565b604082019050919050565b5f6020820190508181035f8301526120bc81612083565b9050919050565b7f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e65725f82015250565b5f6120f7602083611c11565b9150612102826120c3565b602082019050919050565b5f6020820190508181035f830152612124816120eb565b905091905056fea2646970667358221220e13314788996b67aeb787ca9d7627a3e0a78b1d3bd06fe5cb185536f7d34921964736f6c63430008150033\n";
+
+    private static String librariesLinkedBinary;
 
     public static final String FUNC_APPROVEDOPERATORS = "approvedOperators";
 
@@ -108,7 +113,7 @@ public class VideoNftMarketplace_ABI extends Contract {
     ;
 
     @Deprecated
-    protected VideoNftMarketplace_ABI(
+    protected VideoNFTMarketplace(
             String contractAddress,
             Web3j web3j,
             Credentials credentials,
@@ -117,7 +122,7 @@ public class VideoNftMarketplace_ABI extends Contract {
         super(BINARY, contractAddress, web3j, credentials, gasPrice, gasLimit);
     }
 
-    protected VideoNftMarketplace_ABI(
+    protected VideoNFTMarketplace(
             String contractAddress,
             Web3j web3j,
             Credentials credentials,
@@ -126,7 +131,7 @@ public class VideoNftMarketplace_ABI extends Contract {
     }
 
     @Deprecated
-    protected VideoNftMarketplace_ABI(
+    protected VideoNFTMarketplace(
             String contractAddress,
             Web3j web3j,
             TransactionManager transactionManager,
@@ -135,7 +140,7 @@ public class VideoNftMarketplace_ABI extends Contract {
         super(BINARY, contractAddress, web3j, transactionManager, gasPrice, gasLimit);
     }
 
-    protected VideoNftMarketplace_ABI(
+    protected VideoNFTMarketplace(
             String contractAddress,
             Web3j web3j,
             TransactionManager transactionManager,
@@ -461,42 +466,127 @@ public class VideoNftMarketplace_ABI extends Contract {
     }
 
     @Deprecated
-    public static VideoNftMarketplace_ABI load(
+    public static VideoNFTMarketplace load(
             String contractAddress,
             Web3j web3j,
             Credentials credentials,
             BigInteger gasPrice,
             BigInteger gasLimit) {
-        return new VideoNftMarketplace_ABI(contractAddress, web3j, credentials, gasPrice, gasLimit);
+        return new VideoNFTMarketplace(contractAddress, web3j, credentials, gasPrice, gasLimit);
     }
 
     @Deprecated
-    public static VideoNftMarketplace_ABI load(
+    public static VideoNFTMarketplace load(
             String contractAddress,
             Web3j web3j,
             TransactionManager transactionManager,
             BigInteger gasPrice,
             BigInteger gasLimit) {
-        return new VideoNftMarketplace_ABI(
+        return new VideoNFTMarketplace(
                 contractAddress, web3j, transactionManager, gasPrice, gasLimit);
     }
 
-    public static VideoNftMarketplace_ABI load(
+    public static VideoNFTMarketplace load(
             String contractAddress,
             Web3j web3j,
             Credentials credentials,
             ContractGasProvider contractGasProvider) {
-        return new VideoNftMarketplace_ABI(
-                contractAddress, web3j, credentials, contractGasProvider);
+        return new VideoNFTMarketplace(contractAddress, web3j, credentials, contractGasProvider);
     }
 
-    public static VideoNftMarketplace_ABI load(
+    public static VideoNFTMarketplace load(
             String contractAddress,
             Web3j web3j,
             TransactionManager transactionManager,
             ContractGasProvider contractGasProvider) {
-        return new VideoNftMarketplace_ABI(
+        return new VideoNFTMarketplace(
                 contractAddress, web3j, transactionManager, contractGasProvider);
+    }
+
+    public static RemoteCall<VideoNFTMarketplace> deploy(
+            Web3j web3j,
+            Credentials credentials,
+            ContractGasProvider contractGasProvider,
+            String _videoNFTAddress) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, _videoNFTAddress)));
+        return deployRemoteCall(
+                VideoNFTMarketplace.class,
+                web3j,
+                credentials,
+                contractGasProvider,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    public static RemoteCall<VideoNFTMarketplace> deploy(
+            Web3j web3j,
+            TransactionManager transactionManager,
+            ContractGasProvider contractGasProvider,
+            String _videoNFTAddress) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, _videoNFTAddress)));
+        return deployRemoteCall(
+                VideoNFTMarketplace.class,
+                web3j,
+                transactionManager,
+                contractGasProvider,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    @Deprecated
+    public static RemoteCall<VideoNFTMarketplace> deploy(
+            Web3j web3j,
+            Credentials credentials,
+            BigInteger gasPrice,
+            BigInteger gasLimit,
+            String _videoNFTAddress) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, _videoNFTAddress)));
+        return deployRemoteCall(
+                VideoNFTMarketplace.class,
+                web3j,
+                credentials,
+                gasPrice,
+                gasLimit,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    @Deprecated
+    public static RemoteCall<VideoNFTMarketplace> deploy(
+            Web3j web3j,
+            TransactionManager transactionManager,
+            BigInteger gasPrice,
+            BigInteger gasLimit,
+            String _videoNFTAddress) {
+        String encodedConstructor =
+                FunctionEncoder.encodeConstructor(
+                        Arrays.<Type>asList(
+                                new org.web3j.abi.datatypes.Address(160, _videoNFTAddress)));
+        return deployRemoteCall(
+                VideoNFTMarketplace.class,
+                web3j,
+                transactionManager,
+                gasPrice,
+                gasLimit,
+                getDeploymentBinary(),
+                encodedConstructor);
+    }
+
+    //    public static void linkLibraries(List<Contract.LinkReference> references) {
+    //        librariesLinkedBinary = linkBinaryWithReferences(BINARY, references);
+    //    }
+
+    private static String getDeploymentBinary() {
+        return BINARY;
     }
 
     public static class NFTDelistedEventResponse extends BaseEventResponse {
