--- src/main/java/TokenUs/TokenUs_BE/web3/contract/VideoNft_ABI.java
@@ -1,789 +1,0 @@
-package TokenUs.TokenUs_BE.web3.contract;
-
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.Callable;
-
-import io.reactivex.Flowable;
-import org.web3j.abi.EventEncoder;
-import org.web3j.abi.TypeReference;
-import org.web3j.abi.datatypes.Address;
-import org.web3j.abi.datatypes.Bool;
-import org.web3j.abi.datatypes.Event;
-import org.web3j.abi.datatypes.Function;
-import org.web3j.abi.datatypes.Type;
-import org.web3j.abi.datatypes.Utf8String;
-import org.web3j.abi.datatypes.generated.Uint256;
-import org.web3j.crypto.Credentials;
-import org.web3j.protocol.Web3j;
-import org.web3j.protocol.core.DefaultBlockParameter;
-import org.web3j.protocol.core.RemoteFunctionCall;
-import org.web3j.protocol.core.methods.request.EthFilter;
-import org.web3j.protocol.core.methods.response.BaseEventResponse;
-import org.web3j.protocol.core.methods.response.Log;
-import org.web3j.protocol.core.methods.response.TransactionReceipt;
-import org.web3j.tuples.generated.Tuple5;
-import org.web3j.tuples.generated.Tuple6;
-import org.web3j.tx.Contract;
-import org.web3j.tx.TransactionManager;
-import org.web3j.tx.gas.ContractGasProvider;
-
-/**
- * Auto generated code.
- *
- * <p><strong>Do not modify!</strong>
- *
- * <p>Please use the <a href="https://docs.web3j.io/command_line.html">web3j command line tools</a>,
- * or the org.web3j.codegen.SolidityFunctionWrapperGenerator in the <a
- * href="https://github.com/hyperledger-web3j/web3j/tree/main/codegen">codegen module</a> to update.
- *
- * <p>Generated with web3j version 1.6.3.
- */
-@SuppressWarnings("rawtypes")
-public class VideoNft_ABI extends Contract {
-    public static final String BINARY = "Bin file was not provided";
-
-    public static final String FUNC_APPROVE = "approve";
-
-    public static final String FUNC_BALANCEOF = "balanceOf";
-
-    public static final String FUNC_GETAPPROVED = "getApproved";
-
-    public static final String FUNC_ISAPPROVEDFORALL = "isApprovedForAll";
-
-    public static final String FUNC_NAME = "name";
-
-    public static final String FUNC_OWNER = "owner";
-
-    public static final String FUNC_OWNEROF = "ownerOf";
-
-    public static final String FUNC_RENOUNCEOWNERSHIP = "renounceOwnership";
-
-    public static final String FUNC_safeTransferFrom = "safeTransferFrom";
-
-    public static final String FUNC_SETAPPROVALFORALL = "setApprovalForAll";
-
-    public static final String FUNC_SUPPORTSINTERFACE = "supportsInterface";
-
-    public static final String FUNC_SYMBOL = "symbol";
-
-    public static final String FUNC_TOKENBYINDEX = "tokenByIndex";
-
-    public static final String FUNC_TOKENOFOWNERBYINDEX = "tokenOfOwnerByIndex";
-
-    public static final String FUNC_TOKENTOVIDEO = "tokenToVideo";
-
-    public static final String FUNC_TOKENURI = "tokenURI";
-
-    public static final String FUNC_TOTALSUPPLY = "totalSupply";
-
-    public static final String FUNC_TRANSFERFROM = "transferFrom";
-
-    public static final String FUNC_TRANSFEROWNERSHIP = "transferOwnership";
-
-    public static final String FUNC_TRUSTEDOPERATOR = "trustedOperator";
-
-    public static final String FUNC_VIDEOEXISTS = "videoExists";
-
-    public static final String FUNC_VIDEOS = "videos";
-
-    public static final String FUNC_SETTRUSTEDOPERATOR = "setTrustedOperator";
-
-    public static final String FUNC_MINTVIDEONFT = "mintVideoNFT";
-
-    public static final String FUNC_GETVIDEOINFO = "getVideoInfo";
-
-    public static final String FUNC_GETVIDEOIDOFTOKEN = "getVideoIdOfToken";
-
-    public static final Event APPROVAL_EVENT =
-            new Event(
-                    "Approval",
-                    Arrays.<TypeReference<?>>asList(
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Uint256>(true) {}));
-    ;
-
-    public static final Event APPROVALFORALL_EVENT =
-            new Event(
-                    "ApprovalForAll",
-                    Arrays.<TypeReference<?>>asList(
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Bool>() {}));
-    ;
-
-    public static final Event OWNERSHIPTRANSFERRED_EVENT =
-            new Event(
-                    "OwnershipTransferred",
-                    Arrays.<TypeReference<?>>asList(
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Address>(true) {}));
-    ;
-
-    public static final Event TRANSFER_EVENT =
-            new Event(
-                    "Transfer",
-                    Arrays.<TypeReference<?>>asList(
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Uint256>(true) {}));
-    ;
-
-    public static final Event VIDEONFTMINTED_EVENT =
-            new Event(
-                    "VideoNFTMinted",
-                    Arrays.<TypeReference<?>>asList(
-                            new TypeReference<Uint256>(true) {},
-                            new TypeReference<Address>(true) {},
-                            new TypeReference<Uint256>() {},
-                            new TypeReference<Utf8String>() {},
-                            new TypeReference<Utf8String>() {},
-                            new TypeReference<Uint256>() {}));
-    ;
-
-    @Deprecated
-    protected VideoNft_ABI(
-            String contractAddress,
-            Web3j web3j,
-            Credentials credentials,
-            BigInteger gasPrice,
-            BigInteger gasLimit) {
-        super(BINARY, contractAddress, web3j, credentials, gasPrice, gasLimit);
-    }
-
-    protected VideoNft_ABI(
-            String contractAddress,
-            Web3j web3j,
-            Credentials credentials,
-            ContractGasProvider contractGasProvider) {
-        super(BINARY, contractAddress, web3j, credentials, contractGasProvider);
-    }
-
-    @Deprecated
-    protected VideoNft_ABI(
-            String contractAddress,
-            Web3j web3j,
-            TransactionManager transactionManager,
-            BigInteger gasPrice,
-            BigInteger gasLimit) {
-        super(BINARY, contractAddress, web3j, transactionManager, gasPrice, gasLimit);
-    }
-
-    protected VideoNft_ABI(
-            String contractAddress,
-            Web3j web3j,
-            TransactionManager transactionManager,
-            ContractGasProvider contractGasProvider) {
-        super(BINARY, contractAddress, web3j, transactionManager, contractGasProvider);
-    }
-
-    public static List<ApprovalEventResponse> getApprovalEvents(
-            TransactionReceipt transactionReceipt) {
-        List<Contract.EventValuesWithLog> valueList =
-                staticExtractEventParametersWithLog(APPROVAL_EVENT, transactionReceipt);
-        ArrayList<ApprovalEventResponse> responses =
-                new ArrayList<ApprovalEventResponse>(valueList.size());
-        for (Contract.EventValuesWithLog eventValues : valueList) {
-            ApprovalEventResponse typedResponse = new ApprovalEventResponse();
-            typedResponse.log = eventValues.getLog();
-            typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
-            typedResponse.approved = (String) eventValues.getIndexedValues().get(1).getValue();
-            typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
-            responses.add(typedResponse);
-        }
-        return responses;
-    }
-
-    public static ApprovalEventResponse getApprovalEventFromLog(Log log) {
-        Contract.EventValuesWithLog eventValues =
-                staticExtractEventParametersWithLog(APPROVAL_EVENT, log);
-        ApprovalEventResponse typedResponse = new ApprovalEventResponse();
-        typedResponse.log = log;
-        typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
-        typedResponse.approved = (String) eventValues.getIndexedValues().get(1).getValue();
-        typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
-        return typedResponse;
-    }
-
-    public Flowable<ApprovalEventResponse> approvalEventFlowable(EthFilter filter) {
-        return web3j.ethLogFlowable(filter).map(log -> getApprovalEventFromLog(log));
-    }
-
-    public Flowable<ApprovalEventResponse> approvalEventFlowable(
-            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
-        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
-        filter.addSingleTopic(EventEncoder.encode(APPROVAL_EVENT));
-        return approvalEventFlowable(filter);
-    }
-
-    public static List<ApprovalForAllEventResponse> getApprovalForAllEvents(
-            TransactionReceipt transactionReceipt) {
-        List<Contract.EventValuesWithLog> valueList =
-                staticExtractEventParametersWithLog(APPROVALFORALL_EVENT, transactionReceipt);
-        ArrayList<ApprovalForAllEventResponse> responses =
-                new ArrayList<ApprovalForAllEventResponse>(valueList.size());
-        for (Contract.EventValuesWithLog eventValues : valueList) {
-            ApprovalForAllEventResponse typedResponse = new ApprovalForAllEventResponse();
-            typedResponse.log = eventValues.getLog();
-            typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
-            typedResponse.operator = (String) eventValues.getIndexedValues().get(1).getValue();
-            typedResponse.approved = (Boolean) eventValues.getNonIndexedValues().get(0).getValue();
-            responses.add(typedResponse);
-        }
-        return responses;
-    }
-
-    public static ApprovalForAllEventResponse getApprovalForAllEventFromLog(Log log) {
-        Contract.EventValuesWithLog eventValues =
-                staticExtractEventParametersWithLog(APPROVALFORALL_EVENT, log);
-        ApprovalForAllEventResponse typedResponse = new ApprovalForAllEventResponse();
-        typedResponse.log = log;
-        typedResponse.owner = (String) eventValues.getIndexedValues().get(0).getValue();
-        typedResponse.operator = (String) eventValues.getIndexedValues().get(1).getValue();
-        typedResponse.approved = (Boolean) eventValues.getNonIndexedValues().get(0).getValue();
-        return typedResponse;
-    }
-
-    public Flowable<ApprovalForAllEventResponse> approvalForAllEventFlowable(EthFilter filter) {
-        return web3j.ethLogFlowable(filter).map(log -> getApprovalForAllEventFromLog(log));
-    }
-
-    public Flowable<ApprovalForAllEventResponse> approvalForAllEventFlowable(
-            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
-        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
-        filter.addSingleTopic(EventEncoder.encode(APPROVALFORALL_EVENT));
-        return approvalForAllEventFlowable(filter);
-    }
-
-    public static List<OwnershipTransferredEventResponse> getOwnershipTransferredEvents(
-            TransactionReceipt transactionReceipt) {
-        List<Contract.EventValuesWithLog> valueList =
-                staticExtractEventParametersWithLog(OWNERSHIPTRANSFERRED_EVENT, transactionReceipt);
-        ArrayList<OwnershipTransferredEventResponse> responses =
-                new ArrayList<OwnershipTransferredEventResponse>(valueList.size());
-        for (Contract.EventValuesWithLog eventValues : valueList) {
-            OwnershipTransferredEventResponse typedResponse =
-                    new OwnershipTransferredEventResponse();
-            typedResponse.log = eventValues.getLog();
-            typedResponse.previousOwner = (String) eventValues.getIndexedValues().get(0).getValue();
-            typedResponse.newOwner = (String) eventValues.getIndexedValues().get(1).getValue();
-            responses.add(typedResponse);
-        }
-        return responses;
-    }
-
-    public static OwnershipTransferredEventResponse getOwnershipTransferredEventFromLog(Log log) {
-        Contract.EventValuesWithLog eventValues =
-                staticExtractEventParametersWithLog(OWNERSHIPTRANSFERRED_EVENT, log);
-        OwnershipTransferredEventResponse typedResponse = new OwnershipTransferredEventResponse();
-        typedResponse.log = log;
-        typedResponse.previousOwner = (String) eventValues.getIndexedValues().get(0).getValue();
-        typedResponse.newOwner = (String) eventValues.getIndexedValues().get(1).getValue();
-        return typedResponse;
-    }
-
-    public Flowable<OwnershipTransferredEventResponse> ownershipTransferredEventFlowable(
-            EthFilter filter) {
-        return web3j.ethLogFlowable(filter).map(log -> getOwnershipTransferredEventFromLog(log));
-    }
-
-    public Flowable<OwnershipTransferredEventResponse> ownershipTransferredEventFlowable(
-            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
-        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
-        filter.addSingleTopic(EventEncoder.encode(OWNERSHIPTRANSFERRED_EVENT));
-        return ownershipTransferredEventFlowable(filter);
-    }
-
-    public static List<TransferEventResponse> getTransferEvents(
-            TransactionReceipt transactionReceipt) {
-        List<Contract.EventValuesWithLog> valueList =
-                staticExtractEventParametersWithLog(TRANSFER_EVENT, transactionReceipt);
-        ArrayList<TransferEventResponse> responses =
-                new ArrayList<TransferEventResponse>(valueList.size());
-        for (Contract.EventValuesWithLog eventValues : valueList) {
-            TransferEventResponse typedResponse = new TransferEventResponse();
-            typedResponse.log = eventValues.getLog();
-            typedResponse.from = (String) eventValues.getIndexedValues().get(0).getValue();
-            typedResponse.to = (String) eventValues.getIndexedValues().get(1).getValue();
-            typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
-            responses.add(typedResponse);
-        }
-        return responses;
-    }
-
-    public static TransferEventResponse getTransferEventFromLog(Log log) {
-        Contract.EventValuesWithLog eventValues =
-                staticExtractEventParametersWithLog(TRANSFER_EVENT, log);
-        TransferEventResponse typedResponse = new TransferEventResponse();
-        typedResponse.log = log;
-        typedResponse.from = (String) eventValues.getIndexedValues().get(0).getValue();
-        typedResponse.to = (String) eventValues.getIndexedValues().get(1).getValue();
-        typedResponse.tokenId = (BigInteger) eventValues.getIndexedValues().get(2).getValue();
-        return typedResponse;
-    }
-
-    public Flowable<TransferEventResponse> transferEventFlowable(EthFilter filter) {
-        return web3j.ethLogFlowable(filter).map(log -> getTransferEventFromLog(log));
-    }
-
-    public Flowable<TransferEventResponse> transferEventFlowable(
-            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
-        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
-        filter.addSingleTopic(EventEncoder.encode(TRANSFER_EVENT));
-        return transferEventFlowable(filter);
-    }
-
-    public static List<VideoNFTMintedEventResponse> getVideoNFTMintedEvents(
-            TransactionReceipt transactionReceipt) {
-        List<Contract.EventValuesWithLog> valueList =
-                staticExtractEventParametersWithLog(VIDEONFTMINTED_EVENT, transactionReceipt);
-        ArrayList<VideoNFTMintedEventResponse> responses =
-                new ArrayList<VideoNFTMintedEventResponse>(valueList.size());
-        for (Contract.EventValuesWithLog eventValues : valueList) {
-            VideoNFTMintedEventResponse typedResponse = new VideoNFTMintedEventResponse();
-            typedResponse.log = eventValues.getLog();
-            typedResponse.videoId = (BigInteger) eventValues.getIndexedValues().get(0).getValue();
-            typedResponse.creatorAddress =
-                    (String) eventValues.getIndexedValues().get(1).getValue();
-            typedResponse.totalSupply =
-                    (BigInteger) eventValues.getNonIndexedValues().get(0).getValue();
-            typedResponse.name = (String) eventValues.getNonIndexedValues().get(1).getValue();
-            typedResponse.symbol = (String) eventValues.getNonIndexedValues().get(2).getValue();
-            typedResponse.price = (BigInteger) eventValues.getNonIndexedValues().get(3).getValue();
-            responses.add(typedResponse);
-        }
-        return responses;
-    }
-
-    public static VideoNFTMintedEventResponse getVideoNFTMintedEventFromLog(Log log) {
-        Contract.EventValuesWithLog eventValues =
-                staticExtractEventParametersWithLog(VIDEONFTMINTED_EVENT, log);
-        VideoNFTMintedEventResponse typedResponse = new VideoNFTMintedEventResponse();
-        typedResponse.log = log;
-        typedResponse.videoId = (BigInteger) eventValues.getIndexedValues().get(0).getValue();
-        typedResponse.creatorAddress = (String) eventValues.getIndexedValues().get(1).getValue();
-        typedResponse.totalSupply =
-                (BigInteger) eventValues.getNonIndexedValues().get(0).getValue();
-        typedResponse.name = (String) eventValues.getNonIndexedValues().get(1).getValue();
-        typedResponse.symbol = (String) eventValues.getNonIndexedValues().get(2).getValue();
-        typedResponse.price = (BigInteger) eventValues.getNonIndexedValues().get(3).getValue();
-        return typedResponse;
-    }
-
-    public Flowable<VideoNFTMintedEventResponse> videoNFTMintedEventFlowable(EthFilter filter) {
-        return web3j.ethLogFlowable(filter).map(log -> getVideoNFTMintedEventFromLog(log));
-    }
-
-    public Flowable<VideoNFTMintedEventResponse> videoNFTMintedEventFlowable(
-            DefaultBlockParameter startBlock, DefaultBlockParameter endBlock) {
-        EthFilter filter = new EthFilter(startBlock, endBlock, getContractAddress());
-        filter.addSingleTopic(EventEncoder.encode(VIDEONFTMINTED_EVENT));
-        return videoNFTMintedEventFlowable(filter);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> approve(String to, BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_APPROVE,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, to),
-                                new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<BigInteger> balanceOf(String owner) {
-        final Function function =
-                new Function(
-                        FUNC_BALANCEOF,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.Address(160, owner)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
-        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
-    }
-
-    public RemoteFunctionCall<String> getApproved(BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_GETAPPROVED,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<Boolean> isApprovedForAll(String owner, String operator) {
-        final Function function =
-                new Function(
-                        FUNC_ISAPPROVEDFORALL,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, owner),
-                                new org.web3j.abi.datatypes.Address(160, operator)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Bool>() {}));
-        return executeRemoteCallSingleValueReturn(function, Boolean.class);
-    }
-
-    public RemoteFunctionCall<String> name() {
-        final Function function =
-                new Function(
-                        FUNC_NAME,
-                        Arrays.<Type>asList(),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Utf8String>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<String> owner() {
-        final Function function =
-                new Function(
-                        FUNC_OWNER,
-                        Arrays.<Type>asList(),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<String> ownerOf(BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_OWNEROF,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> renounceOwnership() {
-        final Function function =
-                new Function(
-                        FUNC_RENOUNCEOWNERSHIP,
-                        Arrays.<Type>asList(),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> safeTransferFrom(
-            String from, String to, BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_safeTransferFrom,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, from),
-                                new org.web3j.abi.datatypes.Address(160, to),
-                                new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> safeTransferFrom(
-            String from, String to, BigInteger tokenId, byte[] data) {
-        final Function function =
-                new Function(
-                        FUNC_safeTransferFrom,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, from),
-                                new org.web3j.abi.datatypes.Address(160, to),
-                                new org.web3j.abi.datatypes.generated.Uint256(tokenId),
-                                new org.web3j.abi.datatypes.DynamicBytes(data)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> setApprovalForAll(
-            String operator, Boolean approved) {
-        final Function function =
-                new Function(
-                        FUNC_SETAPPROVALFORALL,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, operator),
-                                new org.web3j.abi.datatypes.Bool(approved)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<Boolean> supportsInterface(byte[] interfaceId) {
-        final Function function =
-                new Function(
-                        FUNC_SUPPORTSINTERFACE,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.generated.Bytes4(interfaceId)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Bool>() {}));
-        return executeRemoteCallSingleValueReturn(function, Boolean.class);
-    }
-
-    public RemoteFunctionCall<String> symbol() {
-        final Function function =
-                new Function(
-                        FUNC_SYMBOL,
-                        Arrays.<Type>asList(),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Utf8String>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<BigInteger> tokenByIndex(BigInteger index) {
-        final Function function =
-                new Function(
-                        FUNC_TOKENBYINDEX,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(index)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
-        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
-    }
-
-    public RemoteFunctionCall<BigInteger> tokenOfOwnerByIndex(String owner, BigInteger index) {
-        final Function function =
-                new Function(
-                        FUNC_TOKENOFOWNERBYINDEX,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, owner),
-                                new org.web3j.abi.datatypes.generated.Uint256(index)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
-        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
-    }
-
-    public RemoteFunctionCall<BigInteger> tokenToVideo(BigInteger param0) {
-        final Function function =
-                new Function(
-                        FUNC_TOKENTOVIDEO,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(param0)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
-        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
-    }
-
-    public RemoteFunctionCall<String> tokenURI(BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_TOKENURI,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Utf8String>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<BigInteger> totalSupply() {
-        final Function function =
-                new Function(
-                        FUNC_TOTALSUPPLY,
-                        Arrays.<Type>asList(),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
-        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> transferFrom(
-            String from, String to, BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_TRANSFERFROM,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.Address(160, from),
-                                new org.web3j.abi.datatypes.Address(160, to),
-                                new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> transferOwnership(String newOwner) {
-        final Function function =
-                new Function(
-                        FUNC_TRANSFEROWNERSHIP,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.Address(160, newOwner)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<String> trustedOperator() {
-        final Function function =
-                new Function(
-                        FUNC_TRUSTEDOPERATOR,
-                        Arrays.<Type>asList(),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Address>() {}));
-        return executeRemoteCallSingleValueReturn(function, String.class);
-    }
-
-    public RemoteFunctionCall<Boolean> videoExists(BigInteger param0) {
-        final Function function =
-                new Function(
-                        FUNC_VIDEOEXISTS,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(param0)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Bool>() {}));
-        return executeRemoteCallSingleValueReturn(function, Boolean.class);
-    }
-
-    public RemoteFunctionCall<Tuple6<BigInteger, String, String, String, BigInteger, BigInteger>>
-            videos(BigInteger param0) {
-        final Function function =
-                new Function(
-                        FUNC_VIDEOS,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(param0)),
-                        Arrays.<TypeReference<?>>asList(
-                                new TypeReference<Uint256>() {},
-                                new TypeReference<Utf8String>() {},
-                                new TypeReference<Utf8String>() {},
-                                new TypeReference<Address>() {},
-                                new TypeReference<Uint256>() {},
-                                new TypeReference<Uint256>() {}));
-        return new RemoteFunctionCall<
-                Tuple6<BigInteger, String, String, String, BigInteger, BigInteger>>(
-                function,
-                new Callable<Tuple6<BigInteger, String, String, String, BigInteger, BigInteger>>() {
-                    @Override
-                    public Tuple6<BigInteger, String, String, String, BigInteger, BigInteger> call()
-                            throws Exception {
-                        List<Type> results = executeCallMultipleValueReturn(function);
-                        return new Tuple6<
-                                BigInteger, String, String, String, BigInteger, BigInteger>(
-                                (BigInteger) results.get(0).getValue(),
-                                (String) results.get(1).getValue(),
-                                (String) results.get(2).getValue(),
-                                (String) results.get(3).getValue(),
-                                (BigInteger) results.get(4).getValue(),
-                                (BigInteger) results.get(5).getValue());
-                    }
-                });
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> setTrustedOperator(String _operator) {
-        final Function function =
-                new Function(
-                        FUNC_SETTRUSTEDOPERATOR,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.Address(160, _operator)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<TransactionReceipt> mintVideoNFT(
-            BigInteger videoId,
-            String nftName,
-            String nftSymbol,
-            BigInteger totalSupply,
-            BigInteger price,
-            String creatorAddress) {
-        final Function function =
-                new Function(
-                        FUNC_MINTVIDEONFT,
-                        Arrays.<Type>asList(
-                                new org.web3j.abi.datatypes.generated.Uint256(videoId),
-                                new org.web3j.abi.datatypes.Utf8String(nftName),
-                                new org.web3j.abi.datatypes.Utf8String(nftSymbol),
-                                new org.web3j.abi.datatypes.generated.Uint256(totalSupply),
-                                new org.web3j.abi.datatypes.generated.Uint256(price),
-                                new org.web3j.abi.datatypes.Address(160, creatorAddress)),
-                        Collections.<TypeReference<?>>emptyList());
-        return executeRemoteCallTransaction(function);
-    }
-
-    public RemoteFunctionCall<Tuple5<String, String, String, BigInteger, BigInteger>> getVideoInfo(
-            BigInteger videoId) {
-        final Function function =
-                new Function(
-                        FUNC_GETVIDEOINFO,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(videoId)),
-                        Arrays.<TypeReference<?>>asList(
-                                new TypeReference<Utf8String>() {},
-                                new TypeReference<Utf8String>() {},
-                                new TypeReference<Address>() {},
-                                new TypeReference<Uint256>() {},
-                                new TypeReference<Uint256>() {}));
-        return new RemoteFunctionCall<Tuple5<String, String, String, BigInteger, BigInteger>>(
-                function,
-                new Callable<Tuple5<String, String, String, BigInteger, BigInteger>>() {
-                    @Override
-                    public Tuple5<String, String, String, BigInteger, BigInteger> call()
-                            throws Exception {
-                        List<Type> results = executeCallMultipleValueReturn(function);
-                        return new Tuple5<String, String, String, BigInteger, BigInteger>(
-                                (String) results.get(0).getValue(),
-                                (String) results.get(1).getValue(),
-                                (String) results.get(2).getValue(),
-                                (BigInteger) results.get(3).getValue(),
-                                (BigInteger) results.get(4).getValue());
-                    }
-                });
-    }
-
-    public RemoteFunctionCall<BigInteger> getVideoIdOfToken(BigInteger tokenId) {
-        final Function function =
-                new Function(
-                        FUNC_GETVIDEOIDOFTOKEN,
-                        Arrays.<Type>asList(new org.web3j.abi.datatypes.generated.Uint256(tokenId)),
-                        Arrays.<TypeReference<?>>asList(new TypeReference<Uint256>() {}));
-        return executeRemoteCallSingleValueReturn(function, BigInteger.class);
-    }
-
-    @Deprecated
-    public static VideoNft_ABI load(
-            String contractAddress,
-            Web3j web3j,
-            Credentials credentials,
-            BigInteger gasPrice,
-            BigInteger gasLimit) {
-        return new VideoNft_ABI(contractAddress, web3j, credentials, gasPrice, gasLimit);
-    }
-
-    @Deprecated
-    public static VideoNft_ABI load(
-            String contractAddress,
-            Web3j web3j,
-            TransactionManager transactionManager,
-            BigInteger gasPrice,
-            BigInteger gasLimit) {
-        return new VideoNft_ABI(contractAddress, web3j, transactionManager, gasPrice, gasLimit);
-    }
-
-    public static VideoNft_ABI load(
-            String contractAddress,
-            Web3j web3j,
-            Credentials credentials,
-            ContractGasProvider contractGasProvider) {
-        return new VideoNft_ABI(contractAddress, web3j, credentials, contractGasProvider);
-    }
-
-    public static VideoNft_ABI load(
-            String contractAddress,
-            Web3j web3j,
-            TransactionManager transactionManager,
-            ContractGasProvider contractGasProvider) {
-        return new VideoNft_ABI(contractAddress, web3j, transactionManager, contractGasProvider);
-    }
-
-    public static class ApprovalEventResponse extends BaseEventResponse {
-        public String owner;
-
-        public String approved;
-
-        public BigInteger tokenId;
-    }
-
-    public static class ApprovalForAllEventResponse extends BaseEventResponse {
-        public String owner;
-
-        public String operator;
-
-        public Boolean approved;
-    }
-
-    public static class OwnershipTransferredEventResponse extends BaseEventResponse {
-        public String previousOwner;
-
-        public String newOwner;
-    }
-
-    public static class TransferEventResponse extends BaseEventResponse {
-        public String from;
-
-        public String to;
-
-        public BigInteger tokenId;
-    }
-
-    public static class VideoNFTMintedEventResponse extends BaseEventResponse {
-        public BigInteger videoId;
-
-        public String creatorAddress;
-
-        public BigInteger totalSupply;
-
-        public String name;
-
-        public String symbol;
-
-        public BigInteger price;
-    }
-}
